Array Parameter class

This class is used to represent a parameter has multiple values. Each value is has a label to differentiate values. Only experts should interact directly with this class.

Fields

$idcharacter identifier for parameter.

$namecharacter name of parameter.

$valuevector of values.

$labelcharacter names for each value.

$defaultvector of default values.

$lengthinteger number of values.

$classcharacter class of values.

$rangevector specifying minimum and maximum values.

$widgetfunction used to construct a shiny interface for modifying values.

Usage

x <- ArrayParameter$new(name, value, default, length, 
                                      class, range, widget)
x$print() x$show() x$repr() x$validate(tbl) x$get() x$set(tbl) x$reset() x$render(...)

Arguments

tbl

data.frame containing new parameter values with row names indicating the labels and a column called 'values' containing the new parameter values.

...

arguments passed to function in widget field.

Details

print

print the object.

show

show the object.

repr

character representation of object.

validate

check if a proposed new set of parameters are valid.

get

return a data.frame containing the parameter values.

set

update the parameter values using a data.frame.

reset

update the parameter values to be the default values.

render

create a shiny widget to modify parameter values.

See also

ScalarParameter.