Array Parameter prototype

This prototype is used to represent a parameter has multiple values. Each value is has a label to differentiate values. Only experts should interact directly with this prototype.

Fields

$id

character identifier for parameter.

$name

character name of parameter.

$value

numeric vector of values.

$label

character vector of names for each value.

$default

numeric vector of default values.

$length

integer number of values.

$class

character class of values.

$lower_limit

numeric vector specifying the minimum permitted values.

$upper_limit

numeric vector specifying the maximum permitted values.

$widget

function used to construct a shiny interface for modifying values.

Usage

x$print() x$show() x$repr() x$validate(tbl) x$get() x$set(tbl) x$reset() x$render(...)

Arguments

tbl

data.frame containing new parameter values with row names indicating the labels and a column called "values" containing the new parameter values.

...

arguments passed to function in widget field.

Details

print

print the object.

show

show the object.

repr

character representation of object.

validate

check if a proposed new set of parameters are valid.

get

return a data.frame containing the parameter values.

set

update the parameter values using a data.frame.

reset

update the parameter values to be the default values.

render

create a shiny widget to modify parameter values.

See also

ScalarParameter-class, Parameter-class.