Conservation problem class

This class is used to represent conservation planning problems. A conservation planning problem has spatially explicit planning units. A prioritisation involves making a decision on each planning unit (eg. is the planning unit going to be turned into a protected area?). Each planning unit is associated with a cost that represents the cost incurred by applying the decision to the planning unit. The problem also has a set of representation targets for each feature. Further, it also has constraints used to ensure that the solution meets additional objectives (eg. certain areas are locked into the solution). Finally, a conservation planning problem--unlike an optimization problem--also requires a method to solve the problem. This class represents a planning problem, to actually build and then solve a planning problem, use the problem function. Only experts should use this class directly.

Fields

$datalist object containing data.

$objectiveObjective object used to represent how the targets relate to the solution.

$decisionDecision object used to represent the type of decision made on planning units.

$targetsTarget object used to represent representation targets for features.

$constraintsConstraints object used to represent additional Constraints that the problem is subject to.

$solverSolver object used to solve the problem.

Usage

x <- ConservationProblem$new(cost, features)

x$print() x$show() x$compile() x$solve()

x$get_constraint_parameter(id) x$set_constraint_parameter(id, value) x$render_constraint_parameter(id) x$render_all_constraint_parameters()

x$get_objective_parameter(id) x$set_objective_parameter(id, value) x$render_objective_parameter(id) x$render_all_objective_parameters()

x$get_solver_parameter(id) x$set_solver_parameter(id, value) x$render_solver_parameter(id) x$render_all_solver_parameters()

Arguments

cost

RasterLayer-class, SpatialPolygonsDataFrame, or SpatialLinesDataFrame object showing spatial representation of the planning units and their cost.

features

RasterStack-class object showing distribution of features.

id

id object that refers to a specific parameter.

value

object that the parameter value should become.

Details

print

print the object.

show

show the object.

solve

compile the object and into an OptimizationProblem and then solve it.

get_constraint_parameter

get the value of a parameter (specified by argument id) used in one of the constraints in the object.

set_constraint_parameter

set the value of a parameter (specified by argument id) used in one of the constraints in the object to value.

render_constraint_parameter

generate a shiny widget to modify the the value of a parameter (specified by argument id).

render_all_constraint_parameters

generate a shiny div containing all the parameters' widgets.

get_objective_parameter

get the value of a parameter (specified by argument id) used in one of the objectives in the object.

set_objective_parameter

set the value of a parameter (specified by argument id) used in one of the objectives in the object to value.

render_objective_parameter

generate a shiny widget to modify the the value of a parameter (specified by argument id).

render_all_objective_parameters

generate a shiny div containing all the parameters' widgets.

get_solver_parameter

get the value of a parameter (specified by argument id) used in one of the solvers in the object.

set_solver_parameter

set the value of a parameter (specified by argument id) used in one of the solvers in the object to value.

render_solver_parameter

generate a shiny widget to modify the the value of a parameter (specified by argument id).

render_all_solver_parameters

generate a shiny div containing all the parameters' widgets.