Decision class

Decision class

This class is used to represent the type of decision that is made when prioritizing planning units. This class represents a recipe to make a decision, to actually specify the type of decision in a planning problem, see the help page on decisions. Only experts should use this class directly.

Fields

$namecharacter name of decision.

$vtypecharacter specifying the variable type for planning units. The available types are binary ('B') or semi-continuous ('S').

$parameters\ link{Parameters} object used to customize the decision.

$datalist object with data.

$validatefunction used to validate that the decision type is compatible with a given ConservationProblem object.

$applyfunction used to apply the decision type to a OptimizationProblem object.

Usage

x <- Objective$new(name, vtype, parameters, data,
                                 validate, apply)
x$print() x$show() x$repr()

x$validate(cp) x$apply(op)

x$get_parameter(id) x$get_all_parameters() x$set_parameter(id, value) x$render_parameter(id) x$render_all_parameter()

Arguments

id

id object.

value

any object.

cp

ConservationProblem object.

op

OptimizationProblem object.

Details

print

print the object

show

show the object

validate

validate that data used to formulate the decisions are compatible with data in a ConservationProblem object.

apply

add decision types to an OptimizationProblem object.

get_parameter

retrieve the value of a parameter.

get_all_parameters

generate list containing all the parameters.

set_parameter

change the value of a parameter to new value.

render_parameter

generate a shiny widget to modify the the value of a parameter (specified by argument id).

render_all_parameters

generate a shiny div containing all the parameters' widgets.