Decision class

This class is used to represent the decision variables used in optimization. Only experts should use the fields and methods for this class directly.

See also

Super class

prioritizr::ConservationModifier -> Decision

Methods

Public methods

Decision$apply()

Decision$clone()

Inherited methods

prioritizr::ConservationModifier$calculate()

prioritizr::ConservationModifier$get_data()

prioritizr::ConservationModifier$get_internal()

prioritizr::ConservationModifier$print()

prioritizr::ConservationModifier$repr()

prioritizr::ConservationModifier$set_data()

prioritizr::ConservationModifier$set_internal()

prioritizr::ConservationModifier$show()

Method apply()

Update an optimization problem formulation.

Usage

Decision$apply(x)

Arguments

x

optimization_problem() object.

Returns

Invisible TRUE.

Method clone()

The objects of this class are cloneable with this method.

Usage

Decision$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.