Multi-objective conservation problem class

This class is used to represent multi-objective conservation planning problems. It stores the data (e.g., planning units, and features) and mathematical formulation (e.g., the objective, constraints, and other design criteria) needed to generate prioritizations. Most users should use multi_problem() to generate new multi-objective conservation problem objects, and the functions distributed with the package to interact with them (e.g., number_of_features(), number_of_planning_units()). Only experts should use the fields and methods for this class directly.

See also

Public fields

problems

list containing ConservationProblem objects.

defaults

list indicating if other fields contain defaults.

approach

MultiObjApproach object for specifying the multi-objective optimization appraoch.

solver

Solver object specifying the solver for generating solutions.

Methods

Public methods

MultiObjConservationProblem$new()

MultiObjConservationProblem$summary()

MultiObjConservationProblem$print()

MultiObjConservationProblem$show()

MultiObjConservationProblem$repr()

MultiObjConservationProblem$number_of_planning_units()

MultiObjConservationProblem$is_ids_equivalent_to_indices()

MultiObjConservationProblem$planning_unit_indices()

MultiObjConservationProblem$total_unit_ids()

MultiObjConservationProblem$convert_total_unit_ids_to_indices()

MultiObjConservationProblem$planning_unit_indices_with_finite_costs()

MultiObjConservationProblem$number_of_total_units()

MultiObjConservationProblem$planning_unit_class()

MultiObjConservationProblem$number_of_features()

MultiObjConservationProblem$feature_names()

MultiObjConservationProblem$number_of_problems()

MultiObjConservationProblem$problem_names()

MultiObjConservationProblem$number_of_zones()

MultiObjConservationProblem$zone_names()

MultiObjConservationProblem$add_approach()

MultiObjConservationProblem$add_solver()

MultiObjConservationProblem$clone()

Method new()

Create a new multi-objective conservation problem object.

Usage

MultiObjConservationProblem$new(problems)

Arguments

problems

list containing ConservationProblem objects.

Returns

A new MultiObjConservationProblem object.

Method summary()

Print extended information about the object.

Usage

MultiObjConservationProblem$summary()

Returns

Invisible TRUE.

Method print()

Print concise information about the object.

Usage

MultiObjConservationProblem$print()

Returns

Invisible TRUE.

Method show()

Display concise information about the object.

Usage

MultiObjConservationProblem$show()

Returns

Invisible TRUE.

Method repr()

Generate a character representation of the object.

Usage

MultiObjConservationProblem$repr()

Returns

A character value.

Method number_of_planning_units()

Obtain the number of planning units. The planning units correspond to elements in the cost data (e.g., indices, rows, geometries, cells) that have finite values in at least one zone. In other words, planning unit are elements in the cost data that do not have missing (NA) values in every zone.

Usage

MultiObjConservationProblem$number_of_planning_units()

Returns

An integer value.

Method is_ids_equivalent_to_indices()

Check if planning unit identifiers are equivalent to the planning unit indices? Only FALSE if the planning units are data.frame format.

Usage

MultiObjConservationProblem$is_ids_equivalent_to_indices()

Returns

A logical value.

Method planning_unit_indices()

Obtain the planning unit indices.

Usage

MultiObjConservationProblem$planning_unit_indices()

Returns

An integer vector.

Method total_unit_ids()

Obtain the total unit identifiers.

Usage

MultiObjConservationProblem$total_unit_ids()

Returns

An integer vector.

Method convert_total_unit_ids_to_indices()

Convert total unit identifiers to indices.

Usage

MultiObjConservationProblem$convert_total_unit_ids_to_indices(ids)

Arguments

ids

integer vector with planning unit identifiers.

Returns

An integer vector.

Method planning_unit_indices_with_finite_costs()

Obtain the planning unit indices that are associated with finite cost values.

Usage

MultiObjConservationProblem$planning_unit_indices_with_finite_costs()

Returns

A list of integer vectors. Each list element corresponds to a different zone.

Method number_of_total_units()

Obtain the number of total units. The total units include all elements in the cost data (e.g., indices, rows, geometries, cells), including those with missing (NA) values.

Usage

MultiObjConservationProblem$number_of_total_units()

Returns

An integer value.

Method planning_unit_class()

Get planning unit class.

Usage

MultiObjConservationProblem$planning_unit_class()

Returns

A character value.

Method number_of_features()

Obtain the number of features.

Usage

MultiObjConservationProblem$number_of_features()

Returns

An integer value.

Method feature_names()

Obtain the names of the features.

Usage

MultiObjConservationProblem$feature_names()

Returns

A list of character vectors.

Method number_of_problems()

Obtain the number of problems.

Usage

MultiObjConservationProblem$number_of_problems()

Returns

An integer value.

Method problem_names()

Obtain the names of the problems.

Usage

MultiObjConservationProblem$problem_names()

Returns

A character vector.

Method number_of_zones()

Obtain the number of zones.

Usage

MultiObjConservationProblem$number_of_zones()

Returns

An integer value.

Method zone_names()

Obtain the zone names.

Usage

MultiObjConservationProblem$zone_names()

Returns

A character vector.

Method add_approach()

Create a new object with an approach added to the problem formulation.

Usage

MultiObjConservationProblem$add_approach(x)

Arguments

x

MultiObjApproach object.

Returns

An updated MultiObjConservationProblem object.

Method add_solver()

Create a new object with a solver added to the problem formulation.

Usage

MultiObjConservationProblem$add_solver(x)

Arguments

x

Solver object.

Returns

An updated ConservationProblem object.

Method clone()

The objects of this class are cloneable with this method.

Usage

MultiObjConservationProblem$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.