OptimizationProblem class

This class is used to represent an optimization problem. Data are stored in memory using an external pointer. Only experts interact with this class directly.

Fields

$ptrexternalptr object.

Usage

x <- OptimizationProblem$new()
x$print() x$show() x$repr()

x$ncol() x$nrow() x$ncell()

x$modelsense() x$vtype() x$obj() x$A() x$rhs() x$sense() x$lb() x$ub()

x$pu_indices_obj() x$row_ids() x$column_ids()

Arguments

ptr

externalptr object.

Details

print

print the object.

show

show the object.

repr

character representation of object.

ncol

integer number of columns (variables) in model matrix.

nrow

integer number of rows (constraints) in model matrix.

ncell

integer number of cells in model matrix.

modelsense

character model sense.

vtype

character vector of variable types.

obj

numeric vector of objective function.

A

dgCMatrix-class model matrix

rhs

numeric vector of right-hand-side constraints.

sense

character vector of constraint senses.

lb

numeric vector of lower bounds for each column.

ub

numeric vector of upper bounds for each column.

pu_indices_obj

integer vector of indices in objective function that correspond to the status of planning units.

column_ids

character constraint ids for each column in the model matrix.

row_ids

character constraint ids for each row in the model matrix.