Optimization problem methods

These functions are used to access data from an OptimizationProblem-class object.

nrow(x)

ncol(x)

ncell(x)

modelsense(x)

vtype(x)

obj(x)

A(x)

rhs(x)

sense(x)

lb(x)

ub(x)

number_of_features(x)

number_of_planning_units(x)

col_ids(x)

row_ids(x)

Arguments

x
OptimizationProblem-class object.

Value

dgCMatrix-class, numeric vector, numeric vector, or scalar integer depending on the method used.

Details

The functions return the following data:

nrow

integer number of rows (constraints).

ncol

integer number of columns (decision variables).

ncell

integer number of cells.

modelsense

character describing if the problem is to be maximized ("max") or minimized ("min").

vtype

character describing the type of each decision variable: binary ("B"), semi-continuous ("S"), or continuous ("C")

obj

numeric vector specifying the objective function.

A

dgCMatrix-class matrix object defining the problem matrix.

rhs

numeric vector with right-hand-side linear constraints

sense

character vector with the senses of the linear constraints ("<=", ">=", "=").

lb

numeric lower bound for each decision variable. Missing data values (NA) indicate no lower bound for a given variable.

ub

numeric upper bounds for each decision variable. Missing data values (NA) indicate no upper bound for a given variable.

number_of_planning_units

integer number of planning units in the problem.

number_of_features

integer number of features the problem.