Target class

This class is used to represent targets for optimization. Only experts should use the fields and methods for this class directly.

See also

Super class

prioritizr::ConservationModifier -> Target

Methods

Public methods

Target$output()

Target$clone()

Inherited methods

prioritizr::ConservationModifier$calculate()

prioritizr::ConservationModifier$get_data()

prioritizr::ConservationModifier$get_internal()

prioritizr::ConservationModifier$print()

prioritizr::ConservationModifier$repr()

prioritizr::ConservationModifier$set_data()

prioritizr::ConservationModifier$set_internal()

prioritizr::ConservationModifier$show()

Method output()

Output the targets.

Usage

Target$output()

Returns

tibble::tibble() data frame.

Method clone()

The objects of this class are cloneable with this method.

Usage

Target$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.