Target setting method class

This class is used to represent methods for setting targets. Only experts should use the fields and methods for this class directly.

See also

Public fields

name

character value with name of method.

type

character value denoting the target type.

fun

function for calculating targets.

args

list containing arguments.

frame

defused `call for generating error messages.

Methods

Public methods

TargetMethod$new()

TargetMethod$print()

TargetMethod$calculate_targets()

TargetMethod$calculate_targets_km2()

TargetMethod$calculate_relative_targets()

TargetMethod$calculate_absolute_targets()

TargetMethod$clone()

Method new()

Initialize new object.

Usage

TargetMethod$new(name, type, fun, args, frame)

Arguments

name

character value with name of method.

type

character value denoting the target type. Available options include "relative" and "absolute".

fun

function for calculating targets.

args

list containing arguments.

frame

defused `call for generating error messages.

Returns

A new Method object.

Method print()

Print the object.

Usage

TargetMethod$print(...)

Arguments

...

not used.

Returns

Invisible TRUE.

Method calculate_targets()

Calculate targets expressed in the type of units defined for the method (per $type).

Usage

TargetMethod$calculate_targets(x, features, call = NULL)

Arguments

x

problem() object.

features

integer feature indices.

call

NULL or calling environment.

Returns

A numeric vector with target values.

Method calculate_targets_km2()

Calculate targets as km2.

Usage

TargetMethod$calculate_targets_km2(x, features, call = NULL)

Arguments

x

problem() object.

features

integer feature indices.

call

NULL or calling environment.

Returns

A numeric vector with target values expressed in km2.

Method calculate_relative_targets()

Calculate targets as km2.

Usage

TargetMethod$calculate_relative_targets(x, features, call = NULL)

Arguments

x

problem() object.

features

integer feature indices.

call

NULL or calling environment.

Returns

A numeric vector with target values expressed as relative units.

Method calculate_absolute_targets()

Calculate targets expressed as absolute units.

Usage

TargetMethod$calculate_absolute_targets(x, features, call = NULL)

Arguments

x

problem() object.

features

integer feature indices.

call

NULL or calling environment.

Returns

A numeric vector with target values expressed as absolute units.

Method clone()

The objects of this class are cloneable with this method.

Usage

TargetMethod$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.