Add default decisions

This function adds the default decision types to a conservation planning problem. The default types are binary and are added using the add_binary_decisions function.

add_default_decisions(x)

Arguments

x
ConservationProblem-class object.

See also