Default solver

Identify the best solver currently installed on the system and specify that it should be used to solve a conservation planning problem. Ranked from best to worst, the available solvers that can be used are: Gurobi (add_gurobi_solver), Rsymphony (add_rsymphony_solver), then lpsymphony (add_lpsymphony_solver).

add_default_solver(x, ...)

Arguments

x
ConservationProblem-class object.

...
arguments passed to the solver.

See also