Add targets based on feature groups

Add targets to a conservation planning problem, wherein each feature is assigned to a particular group and a target setting method is specified for each feature group. This function is designed to provide a convenient alternative to add_auto_targets().

Usage

add_group_targets(x, groups, method)

Arguments

x

problem() object.

groups

character vector of group names.

method

list of name-value pairs that describe the target setting method for each group. Here, the names of method correspond to different values in groups. Additionally, the elements of method should be the character names of target setting methods, or objects used to specify target setting methods. See Target methods below for more information.

Value

An updated problem() object with the targets added to it.

Target methods

Here method is used to specify a target setting method for each group. In particular, each element should correspond to a different group, with the name of the element indicating the group and the value denoting the target setting method. One option for specifying a particular target setting method is to use a character value that denotes the name of the method (e.g., "jung" to set targets following Jung et al. 2021). This option is particularly useful if default parameters should be considered during target calculations. Alternatively, another option for specifying a particular target setting method is to use a function to define an object (e.g., jung_target() to set targets following Jung et al. 2021). This alternative option is particularly useful if customized parameters should be considered during target calculations. Note that a method can contain a mix of target setting methods defined using character values and functions.

The following character values can be used to specify target setting methods: "jung" (per Jung et al. 2021), "polak" (per Polak et al. 2016), "rodrigues" (per Rodrigues et al. 2004), "ward" (per Ward et al. 2025), and "watson" (per Watson et al. 2010). (2010). Additionally, the following values can be used to set targets based on criteria from the IUCN Red List of Threatened Species (IUCN 2025): "rl_species_VU_A1_B1" "rl_species_EN_A1_B1", "rl_species_CR_A1_B1", "rl_species_VU_A1_B2" "rl_species_EN_A1_B2", "rl_species_CR_A1_B2", "rl_species_VU_A2_B1" "rl_species_EN_A2_B1", "rl_species_CR_A2_B1", "rl_species_VU_A2_B2" "rl_species_EN_A2_B2", "rl_species_CR_A2_B2", "rl_species_VU_A3_B1" "rl_species_EN_A3_B1", "rl_species_CR_A3_B1", "rl_species_VU_A3_B2" "rl_species_EN_A3_B2", "rl_species_CR_A3_B2", "rl_species_VU_A4_B1" "rl_species_EN_A4_B1", "rl_species_CR_A4_B1", "rl_species_VU_A4_B2" "rl_species_EN_A4_B2", and "rl_species_CR_A4_B2". Furthermore, the following values can be used to set targets based on criteria from the IUCN Red List of Ecosystems (IUCN 2024): "rl_ecosystem_VU_A1_B1", "rl_ecosystem_EN_A1_B1", "rl_ecosystem_CR_A1_B1", "rl_ecosystem_VU_A1_B2", "rl_ecosystem_EN_A1_B2", "rl_ecosystem_CR_A1_B2", "rl_ecosystem_VU_a2a_B1", "rl_ecosystem_EN_a2a_B1", "rl_ecosystem_CR_a2a_B1", "rl_ecosystem_VU_a2a_B2", "rl_ecosystem_EN_a2a_B2", "rl_ecosystem_CR_a2a_B2", "rl_ecosystem_VU_a2b_B1", "rl_ecosystem_EN_a2b_B1", "rl_ecosystem_CR_a2b_B1", "rl_ecosystem_VU_a2b_B2", "rl_ecosystem_EN_a2b_B2", "rl_ecosystem_CR_a2b_B2", "rl_ecosystem_VU_A3_B1", "rl_ecosystem_EN_A3_B1", "rl_ecosystem_CR_A3_B1", "rl_ecosystem_VU_A3_B2", "rl_ecosystem_EN_A3_B2", and "rl_ecosystem_CR_A3_B2".' For convenience, these options can also be specified with lower case letters (e.g., "rl_species_VU_A1_B1" may alternatively be specified as "rl_species_vu_a1_b1"). Note that target setting methods that require additional data or parameters cannot be specified with character values.

The following functions can be used to create an object for specifying a target setting method: spec_absolute_targets(), spec_relative_targets(), spec_interp_absolute_targets(), spec_interp_area_targets(), spec_duran_targets(), spec_jung_targets(), spec_polak_targets(), spec_pop_size_targets(), spec_rl_ecosystem_targets(), spec_rl_species_targets(), spec_rodrigues_targets(), spec_rule_targets(), spec_ward_targets(), spec_watson_targets(), spec_wilson_targets().

Target setting

Many conservation planning problems require targets. Targets are used to specify the minimum amount, or proportion, of a feature's spatial distribution that should ideally be protected. This is important so that the optimization process can weigh the merits and trade-offs between improving the representation of one feature over another feature. Although it can be challenging to set meaningful targets, this is a critical step for ensuring that prioritizations meet the stakeholder objectives that underpin a prioritization exercise (Carwardine et al. 2009). In other words, targets play an important role in ensuring that a priority setting process is properly tuned according to stakeholder requirements. For example, targets provide a mechanism for ensuring that a prioritization secures enough habitat to promote the long-term persistence of each threatened species, culturally important species, or economically important ecosystem service under consideration. Since there is often uncertainty regarding stakeholder objectives (e.g., how much habitat should be protected for a given species) or the influence of particular target on a prioritization (e.g., how would setting a 90% or 100% for a threatened species alter priorities), it is often useful to generate and compare a suite of prioritizations based on different target scenarios.

Many conservation planning problems require targets. Targets are used to specify the minimum amount, or proportion, of a feature's spatial distribution that should ideally be protected. This is important so that the optimization process can weigh the merits and trade-offs between improving the representation of one feature over another feature. Although it can be challenging to set meaningful targets, this is a critical step for ensuring that prioritizations meet the stakeholder objectives that underpin a prioritization exercise (Carwardine et al. 2009). In other words, targets play an important role in ensuring that a priority setting process is properly tuned according to stakeholder requirements. For example, targets provide a mechanism for ensuring that a prioritization secures enough habitat to promote the long-term persistence of each threatened species, culturally important species, or economically important ecosystem service under consideration. Since there is often uncertainty regarding stakeholder objectives (e.g., how much habitat should be protected for a given species) or the influence of particular target on a prioritization (e.g., how would setting a 90% or 100% for a threatened species alter priorities), it is often useful to generate and compare a suite of prioritizations based on different target scenarios.

Data calculations

This function involves calculating targets based on the spatial extent of the features in x. As such, this function may need to make particular assumptions depending on the data used to specify the planning units and features in x. If possible, when building x with problem(), it is recommended to specify features as a terra::rast() object to ensure correctness. Although this function can be used to specify targets for features in other data formats, you must take care to ensure that the information specifying the amount of each feature are in units of 1 km2. To help ensure that you are aware of this limitation, this function will produce a warning to alert you when the features in x belong to such data formats. If you are confident that your data are in the correct units, then you can safely ignore this warning. Below we provide details for ensuring that the targets are calculated correctly.

x has terra::rast() or sf::st_sf() planning units and terra::rast() features

This function will automatically ensure that the targets are calculated correctly. Given this, it is recommended to specify features in either of these when building a problem().

x has sf::st_sf() or data.frame planning units and character features

Here, the argument to features of problem() is used to specify which columns of the planning unit data denote the amount of each feature present within each planning unit. All of these columns must contain values in units of 1 km km2. For example, if features = c("spp1", "spp2", "spp3"), then the columns "spp1", "spp2", and "spp3" of the planning unit denote data the amount of the first, second, and third features (respectively) present within each planning unit. If the column "spp1" has a value of 5 in the first row of the planning unit data, then this means the first planning unit contains 5 km km2 of habitat for the feature "spp1".

x has data.frame planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij must contain a data.frame with the columns "pu", "species", "amount". For a given row, the "pu" column contains an integer value denoting the identifier for the planning unit, the "species" column contains an integer value denoting the identifier for the feature, and "amount" contains a numericvalue denoting the amount of the the feature (per the"species"column) that is present in the planning unit (per"pu"column). Given this, the "amount" column must contain values that in units of 1 km \ifelse{html}{\out{km<sup>2</sup>}}{\eqn{km^2}}. For example, if the argument torijhas a row where the"pu"column has a value of 1, the"species"column has a value of 2, and the"amount"` column has a value of 3: then this means that the planning unit with an identifier of 1 contains 3 km km2 of the feature associated with the identifier of 2.

x has numeric or matrix planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij_matrix of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij_matrix must contain a matrix (or list of matrix objects) that contains numeric values in units of 1 km km2.

This function involves calculating targets based on the spatial extent of the features in x. As such, this function may need to make particular assumptions depending on the data used to specify the planning units and features in x. If possible, when building x with problem(), it is recommended to specify features as a terra::rast() object to ensure correctness. Although this function can be used to specify targets for features in other data formats, you must take care to ensure that the information specifying the amount of each feature are in units of 1 km2. To help ensure that you are aware of this limitation, this function will produce a warning to alert you when the features in x belong to such data formats. If you are confident that your data are in the correct units, then you can safely ignore this warning. Below we provide details for ensuring that the targets are calculated correctly.

x has terra::rast() or sf::st_sf() planning units and terra::rast() features

This function will automatically ensure that the targets are calculated correctly. Given this, it is recommended to specify features in either of these when building a problem().

x has sf::st_sf() or data.frame planning units and character features

Here, the argument to features of problem() is used to specify which columns of the planning unit data denote the amount of each feature present within each planning unit. All of these columns must contain values in units of 1 km km2. For example, if features = c("spp1", "spp2", "spp3"), then the columns "spp1", "spp2", and "spp3" of the planning unit denote data the amount of the first, second, and third features (respectively) present within each planning unit. If the column "spp1" has a value of 5 in the first row of the planning unit data, then this means the first planning unit contains 5 km km2 of habitat for the feature "spp1".

x has data.frame planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij must contain a data.frame with the columns "pu", "species", "amount". For a given row, the "pu" column contains an integer value denoting the identifier for the planning unit, the "species" column contains an integer value denoting the identifier for the feature, and "amount" contains a numericvalue denoting the amount of the the feature (per the"species"column) that is present in the planning unit (per"pu"column). Given this, the "amount" column must contain values that in units of 1 km \ifelse{html}{\out{km<sup>2</sup>}}{\eqn{km^2}}. For example, if the argument torijhas a row where the"pu"column has a value of 1, the"species"column has a value of 2, and the"amount"` column has a value of 3: then this means that the planning unit with an identifier of 1 contains 3 km km2 of the feature associated with the identifier of 2.

x has numeric or matrix planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij_matrix of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij_matrix must contain a matrix (or list of matrix objects) that contains numeric values in units of 1 km km2.

References

Carwardine J, Klein CJ, Wilson KA, Pressey RL, Possingham HP (2009) Hitting the target and missing the point: target‐based conservation planning in context. Conservation Letters, 2: 4–11.

Jung M, Arnell A, de Lamo X, García-Rangel S, Lewis M, Mark J, Merow C, Miles L, Ondo I, Pironon S, Ravilious C, Rivers M, Schepaschenko D, Tallowin O, van Soesbergen A, Govaerts R, Boyle BL, Enquist BJ, Feng X, Gallagher R, Maitner B, Meiri S, Mulligan M, Ofer G, Roll U, Hanson JO, Jetz W, Di Marco M, McGowan J, Rinnan DS, Sachs JD, Lesiv M, Adams VM, Andrew SC, Burger JR, Hannah L, Marquet PA, McCarthy JK, Morueta-Holme N, Newman EA, Park DS, Roehrdanz PR, Svenning J-C, Violle C, Wieringa JJ, Wynne G, Fritz S, Strassburg BBN, Obersteiner M, Kapos V, Burgess N, Schmidt- Traub G, Visconti P (2021) Areas of global importance for conserving terrestrial biodiversity, carbon and water. Nature Ecology and Evolution, 5:1499–1509.

Polak T, Watson JEM, Fuller RA, Joseph LN, Martin TG, Possingham HP, Venter O, Carwardine J (2015) Efficient expansion of global protected areas requires simultaneous planning for species and ecosystems. Royal Society Open Science, 2: 150107.

See also

Examples

#TODO