Boundary matrix

Generate a boundary matrix showing the shared and exposed edges of planning units. This function assumes the data are in a projected coordinate system where distances can be correctly calculated. Thus spatial data in a longitude/latitude coordinate system should be reprojected prior to using this function.

Arguments

x

Raster-class, SpatialLines-class, or SpatialPolygons-class object. Note that if x is a Raster-class object then it must have only one layer.

edges

integer number of edges that each planning unit has when argument to x is a Raster-class object. Defaults to 4. Valid values are 4 and 6.

Value

Matrix{dgTMatrix-class} object.

Details

This function returns a matrix. Cells along the off-diagonal indicating the length of the shared boundary between two planning units. Cells along the diagonal indicate length of a given planning unit's edges that have no neighbors (eg. for edges of planning units found along the coastline).