Convert data frame to matrix

Convert a data frame with row index, column index, and matrix components into a corresponding matrix. By default returns a simple_triplet_matrix sparse matrix from the slam package.

df_to_matrix(x, nrow, ncol, sparse = TRUE, add_lower = FALSE,
  vars = c("i", "j", "v"))

Arguments

x

data.frame; with variables x$i and x$j for the row and column indices, respectively, and x$v for the matrix components. Alternative names for these variables can be specified by vars.

nrow, ncol

integer; the number of rows and columns, respectively. Defaults to the maximum row and column indices, respectively.

sparse

logical; whether to return a sparse matrix or a normal base matrix object.

add_lower

logical; if the data frame represents a symmetric matrix where only the upper triangle is provided explicitly, set add_lower to TRUE to add in these values explicitly to the matrix.

vars

character vector; three element character vector with names of row, column, and matrix component variables repsectively.

Value

A simple_triplet_matrix sparse matrix from the slam package, or a normal base matrix object.

Examples

df <- data.frame(i = c(1, 1, 3), j = c(1, 3, 3), v = c(10, 20, 30))
df_to_matrix(df)

#> A 3x3 simple triplet matrix.

df_to_matrix(df, sparse = FALSE)

#>      [,1] [,2] [,3]
#> [1,]   10    0   20
#> [2,]    0    0    0
#> [3,]    0    0   30

df_to_matrix(df, sparse = FALSE, add_lower = TRUE)

#>      [,1] [,2] [,3]
#> [1,]   10    0   20
#> [2,]    0    0    0
#> [3,]   20    0   30

df <- data.frame(a = c(1, 1), b = c(1, 2), c = c(5, 25))
df_to_matrix(df, sparse = FALSE, vars = c("a", "b", "c"))

#>      [,1] [,2]
#> [1,]    5   25