Locked in constraints

This function adds constraints to lock in planning units to ensure that they are prioritised. For example, it may be desirable to lock in planning units already within protected areas.

locked_in_constraint(x)

Arguments

x

RasterLayer-class, SpatialPolygons, or SpatialPolygons object.

Value

Constraint object.

Details

If a Spatial object is supplied, any planning units that overlap with argument x are locked in. If a RasterLayer-class is supplied, then the raster must contain logical (TRUE or FALSE) values. Planning units that overlap with TRUE cells are locked into the solution.

See also

constraints for a list of all available constraints.

Examples

problem(sim_pu_raster, sim_features) +
  minimum_set_objective() +
  relative_targets(0.2) +
  locked_in(sim_locked_in_raster)

#> Error in UseMethod("problem"): no applicable method for 'problem' applied to an object of class "c('RasterBrick', 'Raster', 'RasterStackBrick', 'BasicRaster')"