Prepare a maximum target coverage prioritization problem

Specify a maximum target coverage systematic conservation prioritization problem from input data in a variety of formats. These are constructor functions for maxtargets_model objects which encapsulate prioritization problems in a standardized format.

maxtargets_model(x, ...)

# S3 method for numeric
maxtargets_model(x, targets, budget, rij,
  locked_in = integer(), locked_out = integer(),
  target_type = c("relative", "absolute"), ...)

# S3 method for Raster
maxtargets_model(x, features, targets, budget, rij,
  locked_in = integer(), locked_out = integer(),
  target_type = c("relative", "absolute"), ...)

# S3 method for SpatialPolygons
maxtargets_model(x, features, targets, budget, rij,
  locked_in = integer(), locked_out = integer(),
  target_type = c("relative", "absolute"), ...)

Arguments

x

object specifying the planning units to use in the reserve design exercise and their corresponding cost. The possible types of objects are: RasterLayer: planning units are the raster cells and cell values should contain the repsective cost. If x is a RasterLayer, it may be desirable to exlcude some planning units from the analysis, for example those outside the study area. To exclude planning units, set the cost for those raster cells to NA. SpatialPolygonsDataFrame: polygons correspond to the planning units and there must be an attribute field named cost. numeric vector: a vector of planning unit costs. In this case the spatial representation of the planning units is not provided. Therefore, the feature argument is ignored and the representation matrix rij is required. MarxanData: a MarxanData object, from the marxan package (available at https://github.com/jeffreyhanson/marxan), specifying a Marxan reserve design problem, which is equivalent to the minimum set cover problem. The marxan package can create MarxanData objects from a variety of data sources, including the standard Marxan input files. Thus this is valuable for those wanted to transition directly from Marxan. Note that all other arguments are ignored if x is a MarxanData object.

...

additional arguments passed on to methods

targets

numeric; representation targets either as proportion (between 0 and 1) of total representation level when target_type = "relative" (the default), or absolute targets when target_type = "absolute". The order of the targets should match the ordering in the features argument.

budget

numeric; budget for reserve.

rij

numeric matrix (optional); matrix of representation levels of conservation features (rows) within planning units (columns). rij can be a sparse matrix from the slam package (i.e. a simple_triplet_matrix) or a normal base matrix object. Alternatively, a data frame representation of this matrix with three variables: feature index (rij$feature), planning unit index (rij$pu), and corresponding representation level (rij$amount). If this matrix is not provided it will be calculated based on the planning units and RasterStack of conservation feature distributions.

locked_in

integer; indices of planning units to lock in to final solution. For example, it may be desirable to lock in planning units already within protected areas. Alternatively, if x is a RasterLayer, then locked_in can be a binary RasterLayer with 1s identifying the planning units to lock in.

locked_out

integer; indices of planning units to lock out of final solution. For example, it may be desirable to lock in planning units that are already heavily developed and therefore have little viable habitat. Alternatively, if x is a RasterLayer, then locked_out can be a binary RasterLayer with 1s identifying the planning units to lock out.

target_type

"relative" or "absolute"; specifies whether the target argument should be interpreted as relative to the total level of representation or as an absolute target

features

RasterStack object; the distribution of conservation features. Missing values (i.e. NAs) can be used to indicate the absence of a feature in a particular cell instead of explicitly setting these cells to zero. If x is a Raster object then features should be defined on the same raster template. If x is a SpatialPolygonsDataFrame features will be summarize over the polygons using summarize_features. Not required if rij is provided.

Value

A maxtargets_model object describing the prioritization problem to be solved. This is an S3 object consisting of a list with the following components:

cost: numeric vector of planning unit costs

rij: representation matrix

targets: absolute feature targets

budget: budget for reserve

locked_in: indices of locked in planning units

locked_out: indices of locked out planning units

included: logical vector indicating which planning units are to be included in the analysis. If all units are to be included, this is single value (TRUE). Using a subset of planning units is only permitted if the x argument is provided as a RasterLayer object.

Details

The maximum target coverage problem is a modified version of the the maximum coverage problem (see maxcover_model). Each conservation feature is assigned a representation target and the objective is to find the set of planning units that meets most targets while remaining within a fixed budget. The cost is often either the area of the planning units or the opportunity cost of foregone commericial activities (e.g. from logging or agriculture). Representation level is typically given by the occupancy within each planning unit, however, some measure of abundance or probability of occurence may also be used. The representation targets ensure that each species is adequately represented in the reserve network.

This problem meant to be a hybrid between the maximum coverage problem and a Marxan-like minimum set cover problem in that it allows for both a budget and targets to be set.

Methods (by class)

numeric: Numeric vector of costs

Raster: RasterLayer of planning units and corresponding costs

SpatialPolygons: SpatialPolygonsData frame of planning units with cost attribute

See also

minsetcover_model for the minimum set cover problem. maxcover_model for the traditional maximum coverage problem.

Examples

# 5x5 raster template
e <- raster::extent(0, 1, 0, 1)
r <- raster::raster(e, nrows = 5, ncols = 5, vals = 1)

# generate 4 random feature distributions
set.seed(419)
f <- raster::stack(r, r, r, r)
f[] <- sample(0:1, raster::nlayers(f) * raster::ncell(f), replace = TRUE)
f <- setNames(f, letters[1:raster::nlayers(f)])
# genrate cost layer
cost <- r
cost[] <- rnorm(raster::ncell(cost), mean = 100, sd = 10)
cost <- setNames(cost, "cost")

# prepare prioritization model with budget at 25% of total cost
# and targets as 50% of the features' distributions
b_25 <- 0.25 * raster::cellStats(cost, "sum")
model <- maxtargets_model(x = cost, features = f, budget = b_25,
                          targets = 0.5)

# the representation matrix (rij) can be supplied explicitly,
# in which case the features argument is no longer required
rep_mat <- unname(t(f[]))
model <- maxtargets_model(x = cost, rij = rep_mat, budget = b_25,
                          targets = 0.5)

# cells can be locked in or out of the final solution
model <- maxtargets_model(x = cost, features = f, budget = b_25,
                          targets = 0.5, locked_in = 6:10,
                          locked_out = 16:20)
# alternatively, binary rasters can be supplied indicating locked in/out
r_locked_in <- raster::raster(r)
r_locked_in[] <- 0
# lock in cells 6-10
r_locked_in[6:10] <- 1
model_raste_lock <- maxtargets_model(x = cost, features = f, budget = b_25,
                                     locked_in = r_locked_in, targets=0.5,
                                     locked_out = 16:20)

# if some cells are to be excluded, e.g. those outside study area, set
# the cost to NA for these cells.
cost_na <- cost
cost_na[6:10] <- NA
model_na <- maxtargets_model(x = cost_na, features = f, budget = b_25,
                             targets = 0.5)
# the model object now contains an included component specifying which
# cells are to be included
model_na$included

#>  [1]  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE
#> [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#> [25]  TRUE

which(!model_na$included)

#> [1]  6  7  8  9 10

# note that the representation matrix now has fewer columns because
# the decision variables corresponding to excluded cells have been removed
dim(model$rij)

#> [1]  4 25

dim(model_na$rij)

#> [1]  4 20

# planning units can also be supplied as a SpatialPolygonsDataFrame object
# with cost stored as an attribute (x$cost). Typically the function takes
# longer to execute with polygons because summarizing features over planning
# units is less efficient.
cost_spdf <- raster::rasterToPolygons(cost)
model_spdf <- maxtargets_model(x = cost_spdf, features = f, budget = b_25,
                               targets = 0.5)