This function set and get the number of threads for processing data when

This function set and get the number of threads for processing data when processing data. Note that this does not influence the number of threads used when solving a conservation problem.

set_number_of_threads(x = 1L)

get_number_of_threads()

Arguments

x

integer number of threads to use for processing.

Value

get_number_of_threads

integer number of threads.

set_number_of_threads

invisible logical indicating success.

Details

To cease processing data in parallel, set the number of threads to one.

See also

is.parallel.