prioritizr

Prioritizr is an R package for solving systematic conservation planning problems. By using integer linear programming (ILP) techniques, this package offers a flexible interface for creating reserve selection problems that can be tailored to the specific needs of the conservation planner (Rodrigues et al. 2000; Billionnet 2013). Conservation problems can be solved using a variety of commercial and open-source exact algorithm solvers. In contrast to the algorithms conventionally used to solve conservation problems, such as greedy heuristics or simulated annealing, the exact algorithms used by prioritizr are guaranteed to find optimal solutions. This package also has the functionality to read Marxan input data (Ball et al. 2009) and find much better solutions in a much shorter period of time than Marxan (Beyer et al. 2016). See the online code repository for more information.

Details

This package contains vignettes that introduce systematic conservation planning and basic usage of the prioritizr package. The supplemental prioritizrdata package contains further example datasets and worked examples. The vignettes in this package are listed below.

Prioritizr Basics

Background information on the concepts and terminology that underpin systematic conservation planning and their usage in this package.

Quickstart Guide

Short walk-through of the prioritizr package with a simulated dataset.

References

Ball IR, Possingham HP, and Watts M (2009) Marxan and relatives: Software for spatial conservation prioritisation in Spatial conservation prioritisation: Quantitative methods and computational tools. Eds Moilanen A, Wilson KA, and Possingham HP. Oxford University Press, Oxford, UK.

Beyer HL, Dujardin Y, Watts ME, and Possingham HP (2016) Solving conservation planning problems with integer linear programming. Ecological Modelling, 228: 14--22.

Billionnet A (2013) Mathematical optimization ideas for biodiversity conservation. European Journal of Operational Research, 231: 514--534.

Rodrigues AS, Cerdeira OJ, and Gaston KJ (2000) Flexibility, efficiency, and accountability: adapting reserve selection algorithms to more complex conservation problems. Ecography, 23: 565--574.