Conservation planning problem

Create a systematic conservation planning problem. This function is used to specify the basic data used in a spatial prioritization problem: the spatial distribution of the planning units and their costs, as well as the features (eg. species, ecosystems) that need to be conserved. After constructing this object, it can be customized to meet specific objectives using targets (see targets) and constraints (see link{constraints}).

problem(x, features, ...)

# S3 method for RasterLater
problem(x, features, ...)

# S3 method for SpatialPolygonsDataFrame
problem(x, features, cost_column = 1, ...)

# S3 method for SpatialLinesDataFrame
problem(x, features, cost_column = 1, ...)

Arguments

x

RasterLayer-class, SpatialPolygonsDataFrame, or SpatialLinesDataFrame object specifying the planning units to use in the reserve design exercise and their corresponding cost. It may be desirable to exlcude some planning units from the analysis, for example those outside the study area. To exclude planning units, set the cost for those raster cells to NA.

features

Raster-class object showing the distribution of conservation features. Missing values (i.e. NAs) can be used to indicate the absence of a feature in a particular cell instead of explicitly setting these cells to zero.

...

additional arguments.

cost_column

character name or integer indicating the column in the attribute table of a Spatial-class object with the cost data.

Value

A ConservationProblem object containing the basic data used to build a prioritization problem.

See also

constraints, objectives, targets.

Examples

# create problem using raster planning unit data
problem(sim_pu_raster, sim_features)

#> Error in UseMethod("problem"): no applicable method for 'problem' applied to an object of class "c('RasterBrick', 'Raster', 'RasterStackBrick', 'BasicRaster')"

# create problem using polygon planning unit data
problem(sim_pu_polygons, sim_features)

#> Error in match.arg(cost_column, names(x)): 'arg' must be NULL or a character vector

# create problem using line planning unit data
problem(sim_pu_lines, sim_features)

#> Error in match.arg(cost_column, names(x)): 'arg' must be NULL or a character vector

# create problem using point planning unit data
problem(sim_pu_points, sim_features)

#> Error in UseMethod("problem"): no applicable method for 'problem' applied to an object of class "c('SpatialPointsDataFrame', 'SpatialPoints', 'Spatial', 'SpatialVector', 'SpatialPointsNULL')"