Feature by planning unit matrix

Generate a matrix showing the amount of each feature in each planning unit (also known as an rij matrix).Each row corresponds to a different feature and each column corresponds to a different feature.

# S4 method for Raster,Raster
rij_matrix(x, y, ...)

# S4 method for Spatial,Raster
rij_matrix(x, y, ...)

Arguments

x
Raster-class or Spatial-class object representing the planning units.

y
Raster-class object representing the features.

...
additional arguments passed to fast_extract if argument to x inherits from a Spatial-class object.

Value

Matrix{dgCMatrix-class} object.

Details

The sparse matrix represents the spatial intersection between the planning units and the features. Rows correspond to planning units, and columns correspond to features. Values correspond to the amount of the feature in the planning unit. For example, the amount of the third species in the second planning unit would be contained in the cell in the third column and in the second column.

This function can take a long to run for big data sets. To reduce processing time, the set_number_of_threads function can be used to allocate more computational resources. Additionally, dealing with planning units represented with SpatialPolygonsDataFrame object, the velox package can be installed to reduce processing time.

Generally, processing Spatial-class data takes much longer to process then Raster-class data, and so it is recommended to use Raster-class data for planning units where possible.

See also

Examples

# load data
data(sim_pu_raster, sim_pu_polygons, sim_pu_zones_stack)

# create rij matrix using raster layer planning units
rij_raster <- rij_matrix(sim_pu_raster, sim_features)
print(rij_raster)

#> 5 x 90 sparse Matrix of class "dgCMatrix"
#>                                                                              
#> layer.1 0.7248183 0.7507224 0.7779643 0.8037375 0.8260278 0.8437103 0.8563937
#> layer.2 0.5574857 0.5181092 0.4838966 0.4562762 0.4350560 0.4187520 0.4051169
#> layer.3 0.6444085 0.6603669 0.6728598 0.6821476 0.6875270 0.6875270 0.6801520
#> layer.4 0.6770378 0.7287698 0.7764622 0.8149431 0.8422364 0.8586300 0.8652172
#> layer.5 0.2362890 0.2402769 0.2508932 0.2656888 0.2822200 0.2984055 0.3129927
#>                                                                              
#> layer.1 0.8641689 0.8674019 0.8666363 0.7511302 0.8016750 0.8267485 0.8653989
#> layer.2 0.3916992 0.3763394 0.3575430 0.5983586 0.5276884 0.5008750 0.4645487
#> layer.3 0.6631214 0.6341720 0.5916197 0.6993602 0.7212457 0.7248101 0.7148262
#> layer.4 0.8628960 0.8519693 0.8321315 0.6795108 0.7704815 0.8073525 0.8512819
#> layer.5 0.3259125 0.3383686 0.3526492 0.2670985 0.2683395 0.2776719 0.3016443
#>                                                                              
#> layer.1 0.8840400 0.8860157 0.8835336 0.7723728 0.7959365 0.8462110 0.8672469
#> layer.2 0.4374908 0.4211294 0.4004226 0.6297735 0.5950759 0.5391960 0.5200131
#> layer.3 0.6714647 0.6321588 0.5791818 0.7450044 0.7562883 0.7605353 0.7533101
#> layer.4 0.8576157 0.8479705 0.8294981 0.6784397 0.7188296 0.7936370 0.8202571
#> layer.5 0.3228155 0.3326603 0.3445537 0.2991644 0.2887639 0.2906407 0.2977391
#>                                                                              
#> layer.1 0.8943194 0.8999777 0.9005832 0.8962624 0.8112065 0.8369930 0.8614053
#> layer.2 0.4925263 0.4789701 0.4620831 0.4401063 0.6207819 0.5929076 0.5700981
#> layer.3 0.7145160 0.6791363 0.6305211 0.5680530 0.7888894 0.7915654 0.7876625
#> layer.4 0.8465113 0.8467459 0.8386354 0.8217711 0.7059923 0.7406859 0.7725552
#> layer.5 0.3137670 0.3209210 0.3283011 0.3379912 0.3103051 0.3024799 0.3012667
#>                                                                              
#> layer.1 0.8819718 0.8974209 0.9074066 0.9120139 0.9113361 0.9052485 0.7957064
#> layer.2 0.5525281 0.5388267 0.5265852 0.5130207 0.4956148 0.4726128 0.6650257
#> layer.3 0.7766139 0.7571669 0.7274801 0.6854177 0.6292798 0.5591331 0.8029006
#> layer.4 0.7983796 0.8165496 0.8266075 0.8285589 0.8223247 0.8074381 0.6627752
#> layer.5 0.3041022 0.3087694 0.3137525 0.3185802 0.3240150 0.3319974 0.3472246
#>                                                                              
#> layer.1 0.8206631 0.8472434 0.8923624 0.9073109 0.9165886 0.9203324 0.9185496
#> layer.2 0.6379358 0.6133510 0.5768621 0.5638298 0.5515773 0.5374756 0.5191493
#> layer.3 0.8108013 0.8117542 0.7922287 0.7696561 0.7361006 0.6893752 0.6279826
#> layer.4 0.6871299 0.7151948 0.7672320 0.7856588 0.7971341 0.8011305 0.7972704
#> layer.5 0.3248820 0.3123190 0.3065302 0.3085559 0.3115342 0.3148613 0.3191195
#>                                                                              
#> layer.1 0.9108490 0.7968338 0.8240176 0.8521592 0.8777354 0.8984012 0.9131857
#> layer.2 0.4949689 0.6695972 0.6467601 0.6256222 0.6075654 0.5926670 0.5797533
#> layer.3 0.5526364 0.8164170 0.8229822 0.8225931 0.8151675 0.7999022 0.7753104
#> layer.4 0.7849815 0.6473763 0.6621084 0.6821691 0.7044423 0.7258117 0.7437278
#> layer.5 0.3260539 0.3561070 0.3303721 0.3146897 0.3067608 0.3041676 0.3047138
#>                                                                              
#> layer.1 0.9220332 0.9251440 0.9132930 0.7908189 0.8212131 0.8516865 0.8787320
#> layer.2 0.5667843 0.5514072 0.5057420 0.6659953 0.6476484 0.6300742 0.6142849
#> layer.3 0.7393581 0.6899044 0.5479994 0.8212380 0.8264482 0.8247911 0.8160955
#> layer.4 0.7564196 0.7627596 0.7532268 0.6269176 0.6313261 0.6421905 0.6573929
#> layer.5 0.3067629 0.3095698 0.3200481 0.3536143 0.3265738 0.3096008 0.3006110
#>                                                                              
#> layer.1 0.9001480 0.9151646 0.9266009 0.9231565 0.9126784 0.7780055 0.8124351
#> layer.2 0.6002388 0.5868879 0.5550737 0.5329318 0.5050504 0.6548266 0.6412811
#> layer.3 0.7995157 0.7735758 0.6858468 0.6212074 0.5440173 0.8183262 0.8219222
#> layer.4 0.6743762 0.6906257 0.7128474 0.7158740 0.7119946 0.6017931 0.5957785
#> layer.5 0.2973334 0.2976404 0.3031043 0.3074833 0.3141743 0.3413477 0.3149651
#>                                                                              
#> layer.1 0.8459157 0.8750037 0.8976629 0.9133201 0.9222403 0.9247606 0.9207040
#> layer.2 0.6274436 0.6139020 0.6004693 0.5862360 0.5698132 0.5497017 0.5247025
#> layer.3 0.8187907 0.8087143 0.7908843 0.7639590 0.7262465 0.6761764 0.6132236
#> layer.4 0.5968458 0.6038925 0.6151398 0.6284744 0.6417581 0.6530286 0.6605704
#> layer.5 0.2984347 0.2898809 0.2872081 0.2883890 0.2917381 0.2962070 0.3016299
#>                                                                              
#> layer.1 0.7592975 0.7982456 0.8351602 0.8667091 0.8910204 0.9076832 0.9170826
#> layer.2 0.6371475 0.6287849 0.6189274 0.6077059 0.5947743 0.5793717 0.5605056
#> layer.3 0.8083373 0.8098167 0.8047655 0.7929835 0.7737941 0.7461143 0.7086380
#> layer.4 0.5726722 0.5569699 0.5485988 0.5473588 0.5522837 0.5618777 0.5743396
#> layer.5 0.3222660 0.2981015 0.2833854 0.2764416 0.2753692 0.2782657 0.2834456
#>                                                                              
#> layer.1 0.9196108 0.9150695 0.9022957 0.7363809 0.7797827 0.8201285 0.8542512
#> layer.2 0.5372286 0.5089562 0.4757551 0.6146763 0.6118445 0.6062060 0.5974179
#> layer.3 0.6602567 0.6008185 0.5321080 0.7917722 0.7904418 0.7828707 0.7689694
#> layer.4 0.5877396 0.6001418 0.6096885 0.5404066 0.5166507 0.5002732 0.4917840
#> layer.5 0.2897038 0.2965682 0.3044346 0.2998922 0.2788979 0.2668354 0.2622215
#>                                                                    
#> layer.1 0.8804256 0.8983432 0.9084501 0.9111669 0.9062951 0.8926558
#> layer.2 0.5849708 0.5682593 0.5467184 0.5200165 0.4882735 0.4522344
#> layer.3 0.7483032 0.7201520 0.6836592 0.6381597 0.5837481 0.5219951
#> layer.4 0.4909064 0.4967779 0.5081142 0.5233099 0.5405023 0.5576481
#> layer.5 0.2633591 0.2684843 0.2759223 0.2843119 0.2928587 0.3015145

# create rij matrix using polygon planning units
rij_polygons <- rij_matrix(sim_pu_polygons, sim_features)
print(rij_polygons)

#> 5 x 90 sparse Matrix of class "dgCMatrix"
#>                                                                              
#> layer.1 0.7248183 0.7507224 0.7779643 0.8037375 0.8260278 0.8437103 0.8563937
#> layer.2 0.5574857 0.5181092 0.4838966 0.4562762 0.4350560 0.4187520 0.4051169
#> layer.3 0.6444085 0.6603669 0.6728598 0.6821476 0.6875270 0.6875270 0.6801520
#> layer.4 0.6770378 0.7287698 0.7764622 0.8149431 0.8422364 0.8586300 0.8652172
#> layer.5 0.2362890 0.2402769 0.2508932 0.2656888 0.2822200 0.2984055 0.3129927
#>                                                                              
#> layer.1 0.8641689 0.8674019 0.8666363 0.7511302 0.8016750 0.8267485 0.8653989
#> layer.2 0.3916992 0.3763394 0.3575430 0.5983586 0.5276884 0.5008750 0.4645487
#> layer.3 0.6631214 0.6341720 0.5916197 0.6993602 0.7212457 0.7248101 0.7148262
#> layer.4 0.8628960 0.8519693 0.8321315 0.6795108 0.7704815 0.8073525 0.8512819
#> layer.5 0.3259125 0.3383686 0.3526492 0.2670985 0.2683395 0.2776719 0.3016443
#>                                                                              
#> layer.1 0.8840400 0.8860157 0.8835336 0.7723728 0.7959365 0.8462110 0.8672469
#> layer.2 0.4374908 0.4211294 0.4004226 0.6297735 0.5950759 0.5391960 0.5200131
#> layer.3 0.6714647 0.6321588 0.5791818 0.7450044 0.7562883 0.7605353 0.7533101
#> layer.4 0.8576157 0.8479705 0.8294981 0.6784397 0.7188296 0.7936370 0.8202571
#> layer.5 0.3228155 0.3326603 0.3445537 0.2991644 0.2887639 0.2906407 0.2977391
#>                                                                              
#> layer.1 0.8943194 0.8999777 0.9005832 0.8962624 0.8112065 0.8369930 0.8614053
#> layer.2 0.4925263 0.4789701 0.4620831 0.4401063 0.6207819 0.5929076 0.5700981
#> layer.3 0.7145160 0.6791363 0.6305211 0.5680530 0.7888894 0.7915654 0.7876625
#> layer.4 0.8465113 0.8467459 0.8386354 0.8217711 0.7059923 0.7406859 0.7725552
#> layer.5 0.3137670 0.3209210 0.3283011 0.3379912 0.3103051 0.3024799 0.3012667
#>                                                                              
#> layer.1 0.8819718 0.8974209 0.9074066 0.9120139 0.9113361 0.9052485 0.7957064
#> layer.2 0.5525281 0.5388267 0.5265852 0.5130207 0.4956148 0.4726128 0.6650257
#> layer.3 0.7766139 0.7571669 0.7274801 0.6854177 0.6292798 0.5591331 0.8029006
#> layer.4 0.7983796 0.8165496 0.8266075 0.8285589 0.8223247 0.8074381 0.6627752
#> layer.5 0.3041022 0.3087694 0.3137525 0.3185802 0.3240150 0.3319974 0.3472246
#>                                                                              
#> layer.1 0.8206631 0.8472434 0.8923624 0.9073109 0.9165886 0.9203324 0.9185496
#> layer.2 0.6379358 0.6133510 0.5768621 0.5638298 0.5515773 0.5374756 0.5191493
#> layer.3 0.8108013 0.8117542 0.7922287 0.7696561 0.7361006 0.6893752 0.6279826
#> layer.4 0.6871299 0.7151948 0.7672320 0.7856588 0.7971341 0.8011305 0.7972704
#> layer.5 0.3248820 0.3123190 0.3065302 0.3085559 0.3115342 0.3148613 0.3191195
#>                                                                              
#> layer.1 0.9108490 0.7968338 0.8240176 0.8521592 0.8777354 0.8984012 0.9131857
#> layer.2 0.4949689 0.6695972 0.6467601 0.6256222 0.6075654 0.5926670 0.5797533
#> layer.3 0.5526364 0.8164170 0.8229822 0.8225931 0.8151675 0.7999022 0.7753104
#> layer.4 0.7849815 0.6473763 0.6621084 0.6821691 0.7044423 0.7258117 0.7437278
#> layer.5 0.3260539 0.3561070 0.3303721 0.3146897 0.3067608 0.3041676 0.3047138
#>                                                                              
#> layer.1 0.9220332 0.9251440 0.9132930 0.7908189 0.8212131 0.8516865 0.8787320
#> layer.2 0.5667843 0.5514072 0.5057420 0.6659953 0.6476484 0.6300742 0.6142849
#> layer.3 0.7393581 0.6899044 0.5479994 0.8212380 0.8264482 0.8247911 0.8160955
#> layer.4 0.7564196 0.7627596 0.7532268 0.6269176 0.6313261 0.6421905 0.6573929
#> layer.5 0.3067629 0.3095698 0.3200481 0.3536143 0.3265738 0.3096008 0.3006110
#>                                                                              
#> layer.1 0.9001480 0.9151646 0.9266009 0.9231565 0.9126784 0.7780055 0.8124351
#> layer.2 0.6002388 0.5868879 0.5550737 0.5329318 0.5050504 0.6548266 0.6412811
#> layer.3 0.7995157 0.7735758 0.6858468 0.6212074 0.5440173 0.8183262 0.8219222
#> layer.4 0.6743762 0.6906257 0.7128474 0.7158740 0.7119946 0.6017931 0.5957785
#> layer.5 0.2973334 0.2976404 0.3031043 0.3074833 0.3141743 0.3413477 0.3149651
#>                                                                              
#> layer.1 0.8459157 0.8750037 0.8976629 0.9133201 0.9222403 0.9247606 0.9207040
#> layer.2 0.6274436 0.6139020 0.6004693 0.5862360 0.5698132 0.5497017 0.5247025
#> layer.3 0.8187907 0.8087143 0.7908843 0.7639590 0.7262465 0.6761764 0.6132236
#> layer.4 0.5968458 0.6038925 0.6151398 0.6284744 0.6417581 0.6530286 0.6605704
#> layer.5 0.2984347 0.2898809 0.2872081 0.2883890 0.2917381 0.2962070 0.3016299
#>                                                                              
#> layer.1 0.7592975 0.7982456 0.8351602 0.8667091 0.8910204 0.9076832 0.9170826
#> layer.2 0.6371475 0.6287849 0.6189274 0.6077059 0.5947743 0.5793717 0.5605056
#> layer.3 0.8083373 0.8098167 0.8047655 0.7929835 0.7737941 0.7461143 0.7086380
#> layer.4 0.5726722 0.5569699 0.5485988 0.5473588 0.5522837 0.5618777 0.5743396
#> layer.5 0.3222660 0.2981015 0.2833854 0.2764416 0.2753692 0.2782657 0.2834456
#>                                                                              
#> layer.1 0.9196108 0.9150695 0.9022957 0.7363809 0.7797827 0.8201285 0.8542512
#> layer.2 0.5372286 0.5089562 0.4757551 0.6146763 0.6118445 0.6062060 0.5974179
#> layer.3 0.6602567 0.6008185 0.5321080 0.7917722 0.7904418 0.7828707 0.7689694
#> layer.4 0.5877396 0.6001418 0.6096885 0.5404066 0.5166507 0.5002732 0.4917840
#> layer.5 0.2897038 0.2965682 0.3044346 0.2998922 0.2788979 0.2668354 0.2622215
#>                                                                    
#> layer.1 0.8804256 0.8983432 0.9084501 0.9111669 0.9062951 0.8926558
#> layer.2 0.5849708 0.5682593 0.5467184 0.5200165 0.4882735 0.4522344
#> layer.3 0.7483032 0.7201520 0.6836592 0.6381597 0.5837481 0.5219951
#> layer.4 0.4909064 0.4967779 0.5081142 0.5233099 0.5405023 0.5576481
#> layer.5 0.2633591 0.2684843 0.2759223 0.2843119 0.2928587 0.3015145

# create rij matrix using raster stack planning units
rij_raster <- rij_matrix(sim_pu_zones_stack, sim_features)
print(rij_raster)

#> 5 x 90 sparse Matrix of class "dgCMatrix"
#>                                                                              
#> layer.1 0.7248183 0.7507224 0.7779643 0.8037375 0.8260278 0.8437103 0.8563937
#> layer.2 0.5574857 0.5181092 0.4838966 0.4562762 0.4350560 0.4187520 0.4051169
#> layer.3 0.6444085 0.6603669 0.6728598 0.6821476 0.6875270 0.6875270 0.6801520
#> layer.4 0.6770378 0.7287698 0.7764622 0.8149431 0.8422364 0.8586300 0.8652172
#> layer.5 0.2362890 0.2402769 0.2508932 0.2656888 0.2822200 0.2984055 0.3129927
#>                                                                              
#> layer.1 0.8641689 0.8674019 0.8666363 0.7511302 0.8016750 0.8267485 0.8653989
#> layer.2 0.3916992 0.3763394 0.3575430 0.5983586 0.5276884 0.5008750 0.4645487
#> layer.3 0.6631214 0.6341720 0.5916197 0.6993602 0.7212457 0.7248101 0.7148262
#> layer.4 0.8628960 0.8519693 0.8321315 0.6795108 0.7704815 0.8073525 0.8512819
#> layer.5 0.3259125 0.3383686 0.3526492 0.2670985 0.2683395 0.2776719 0.3016443
#>                                                                              
#> layer.1 0.8840400 0.8860157 0.8835336 0.7723728 0.7959365 0.8462110 0.8672469
#> layer.2 0.4374908 0.4211294 0.4004226 0.6297735 0.5950759 0.5391960 0.5200131
#> layer.3 0.6714647 0.6321588 0.5791818 0.7450044 0.7562883 0.7605353 0.7533101
#> layer.4 0.8576157 0.8479705 0.8294981 0.6784397 0.7188296 0.7936370 0.8202571
#> layer.5 0.3228155 0.3326603 0.3445537 0.2991644 0.2887639 0.2906407 0.2977391
#>                                                                              
#> layer.1 0.8943194 0.8999777 0.9005832 0.8962624 0.8112065 0.8369930 0.8614053
#> layer.2 0.4925263 0.4789701 0.4620831 0.4401063 0.6207819 0.5929076 0.5700981
#> layer.3 0.7145160 0.6791363 0.6305211 0.5680530 0.7888894 0.7915654 0.7876625
#> layer.4 0.8465113 0.8467459 0.8386354 0.8217711 0.7059923 0.7406859 0.7725552
#> layer.5 0.3137670 0.3209210 0.3283011 0.3379912 0.3103051 0.3024799 0.3012667
#>                                                                              
#> layer.1 0.8819718 0.8974209 0.9074066 0.9120139 0.9113361 0.9052485 0.7957064
#> layer.2 0.5525281 0.5388267 0.5265852 0.5130207 0.4956148 0.4726128 0.6650257
#> layer.3 0.7766139 0.7571669 0.7274801 0.6854177 0.6292798 0.5591331 0.8029006
#> layer.4 0.7983796 0.8165496 0.8266075 0.8285589 0.8223247 0.8074381 0.6627752
#> layer.5 0.3041022 0.3087694 0.3137525 0.3185802 0.3240150 0.3319974 0.3472246
#>                                                                              
#> layer.1 0.8206631 0.8472434 0.8923624 0.9073109 0.9165886 0.9203324 0.9185496
#> layer.2 0.6379358 0.6133510 0.5768621 0.5638298 0.5515773 0.5374756 0.5191493
#> layer.3 0.8108013 0.8117542 0.7922287 0.7696561 0.7361006 0.6893752 0.6279826
#> layer.4 0.6871299 0.7151948 0.7672320 0.7856588 0.7971341 0.8011305 0.7972704
#> layer.5 0.3248820 0.3123190 0.3065302 0.3085559 0.3115342 0.3148613 0.3191195
#>                                                                              
#> layer.1 0.9108490 0.7968338 0.8240176 0.8521592 0.8777354 0.8984012 0.9131857
#> layer.2 0.4949689 0.6695972 0.6467601 0.6256222 0.6075654 0.5926670 0.5797533
#> layer.3 0.5526364 0.8164170 0.8229822 0.8225931 0.8151675 0.7999022 0.7753104
#> layer.4 0.7849815 0.6473763 0.6621084 0.6821691 0.7044423 0.7258117 0.7437278
#> layer.5 0.3260539 0.3561070 0.3303721 0.3146897 0.3067608 0.3041676 0.3047138
#>                                                                              
#> layer.1 0.9220332 0.9251440 0.9132930 0.7908189 0.8212131 0.8516865 0.8787320
#> layer.2 0.5667843 0.5514072 0.5057420 0.6659953 0.6476484 0.6300742 0.6142849
#> layer.3 0.7393581 0.6899044 0.5479994 0.8212380 0.8264482 0.8247911 0.8160955
#> layer.4 0.7564196 0.7627596 0.7532268 0.6269176 0.6313261 0.6421905 0.6573929
#> layer.5 0.3067629 0.3095698 0.3200481 0.3536143 0.3265738 0.3096008 0.3006110
#>                                                                              
#> layer.1 0.9001480 0.9151646 0.9266009 0.9231565 0.9126784 0.7780055 0.8124351
#> layer.2 0.6002388 0.5868879 0.5550737 0.5329318 0.5050504 0.6548266 0.6412811
#> layer.3 0.7995157 0.7735758 0.6858468 0.6212074 0.5440173 0.8183262 0.8219222
#> layer.4 0.6743762 0.6906257 0.7128474 0.7158740 0.7119946 0.6017931 0.5957785
#> layer.5 0.2973334 0.2976404 0.3031043 0.3074833 0.3141743 0.3413477 0.3149651
#>                                                                              
#> layer.1 0.8459157 0.8750037 0.8976629 0.9133201 0.9222403 0.9247606 0.9207040
#> layer.2 0.6274436 0.6139020 0.6004693 0.5862360 0.5698132 0.5497017 0.5247025
#> layer.3 0.8187907 0.8087143 0.7908843 0.7639590 0.7262465 0.6761764 0.6132236
#> layer.4 0.5968458 0.6038925 0.6151398 0.6284744 0.6417581 0.6530286 0.6605704
#> layer.5 0.2984347 0.2898809 0.2872081 0.2883890 0.2917381 0.2962070 0.3016299
#>                                                                              
#> layer.1 0.7592975 0.7982456 0.8351602 0.8667091 0.8910204 0.9076832 0.9170826
#> layer.2 0.6371475 0.6287849 0.6189274 0.6077059 0.5947743 0.5793717 0.5605056
#> layer.3 0.8083373 0.8098167 0.8047655 0.7929835 0.7737941 0.7461143 0.7086380
#> layer.4 0.5726722 0.5569699 0.5485988 0.5473588 0.5522837 0.5618777 0.5743396
#> layer.5 0.3222660 0.2981015 0.2833854 0.2764416 0.2753692 0.2782657 0.2834456
#>                                                                              
#> layer.1 0.9196108 0.9150695 0.9022957 0.7363809 0.7797827 0.8201285 0.8542512
#> layer.2 0.5372286 0.5089562 0.4757551 0.6146763 0.6118445 0.6062060 0.5974179
#> layer.3 0.6602567 0.6008185 0.5321080 0.7917722 0.7904418 0.7828707 0.7689694
#> layer.4 0.5877396 0.6001418 0.6096885 0.5404066 0.5166507 0.5002732 0.4917840
#> layer.5 0.2897038 0.2965682 0.3044346 0.2998922 0.2788979 0.2668354 0.2622215
#>                                                                    
#> layer.1 0.8804256 0.8983432 0.9084501 0.9111669 0.9062951 0.8926558
#> layer.2 0.5849708 0.5682593 0.5467184 0.5200165 0.4882735 0.4522344
#> layer.3 0.7483032 0.7201520 0.6836592 0.6381597 0.5837481 0.5219951
#> layer.4 0.4909064 0.4967779 0.5081142 0.5233099 0.5405023 0.5576481
#> layer.5 0.2633591 0.2684843 0.2759223 0.2843119 0.2928587 0.3015145