Summarize results of prioritization exercise

Summarize results of prioritization exercise

solution_summary(model, results)

Arguments

model

prioritizr_model object

results

prioritizr_results object

Value

A list containing the number of planning units (n_pu), total cost (cost), representation level of each feature (amount_held), and whether or not each target was achieved, for the prioritization solution.

Examples

# raster 100x100 template
e <- raster::extent(0, 100, 0, 100)
r <- raster::raster(e, nrows = 100, ncols = 100, vals = 1)

# generate 9 feature distributions with different scales and range sizes
f <- mapply(function(x, y, r) gaussian_field(r = r, range = x, prop = y),
            rep(c(5, 15, 25), each = 3),
            rep(c(0.1, 0.25, 0.5), times = 3),
            MoreArgs = list(r = r))
f <- raster::stack(f)
f <- setNames(f, letters[1:raster::nlayers(f)])
# genrate cost layer
cost <- gaussian_field(r, 20, mean = 1000, variance = 500)
cost <- setNames(cost, "cost")

# prepare minimum set cover prioritization model
# use 20% targets
msc_model <- minsetcover_model(x = cost, features = f, targets = 0.2)

# solve
msc_results <- prioritize(msc_model)

#> Optimize a model with 9 rows, 10000 columns and 25500 nonzeros
#> Coefficient statistics:
#>   Matrix range    [1e+00, 1e+00]
#>   Objective range [9e+02, 1e+03]
#>   Bounds range    [1e+00, 1e+00]
#>   RHS range       [2e+02, 1e+03]
#> Found heuristic solution: objective 1.54557e+06
#> Presolve removed 0 rows and 329 columns
#> Presolve time: 0.05s
#> Presolved: 9 rows, 9671 columns, 25500 nonzeros
#> Variable types: 0 continuous, 9671 integer (9671 binary)
#> 
#> Root relaxation: objective 1.076168e+06, 49 iterations, 0.02 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1076168.16    0    8 1545572.35 1076168.16  30.4%     -    0s
#> H    0     0                    1076722.4667 1076168.16  0.05%     -    0s
#> H    0     0                    1076689.1599 1076168.16  0.05%     -    0s
#> H    0     0                    1076593.2447 1076168.16  0.04%     -    0s
#>      0     0 1076168.19    0   10 1076593.24 1076168.19  0.04%     -    0s
#>      0     0 1076168.19    0    8 1076593.24 1076168.19  0.04%     -    0s
#>      0     0 1076168.19    0   10 1076593.24 1076168.19  0.04%     -    0s
#>      0     0 1076168.28    0   10 1076593.24 1076168.28  0.04%     -    0s
#>      0     2 1076168.28    0   10 1076593.24 1076168.28  0.04%     -    0s
#> H  104   105                    1076501.3549 1076169.48  0.03%   2.0    1s
#> H 1513  1338                    1076489.8292 1076169.48  0.03%   1.3    4s
#>   1541  1372 1076330.97  300   10 1076489.83 1076169.58  0.03%   1.3    5s
#> H 1545  1306                    1076411.5381 1076252.33  0.01%   1.3    5s
#> H 1548  1243                    1076410.7223 1076252.55  0.01%   1.4    5s
#> H 1550  1183                    1076350.4871 1076252.55  0.01%   1.4    5s
#> 
#> Explored 1550 nodes (2297 simplex iterations) in 5.67 seconds
#> Thread count was 2 (of 2 available processors)
#> 
#> Optimal solution found (tolerance 1.00e-04)
#> Best objective 1.076350487128e+06, best bound 1.076252545658e+06, gap 0.0091%

summary(msc_model, msc_results)

#>            Length Class                 Mode   
#> cost       10000  -none-                numeric
#> rij            6  simple_triplet_matrix list   
#> targets        9  -none-                numeric
#> locked_in      0  -none-                numeric
#> locked_out     0  -none-                numeric
#> included       1  -none-                logical