Solve a problem

Solve a ConservationProblem or an OptimizationProblem.

Arguments

a

ConservationProblem or an OptimizationProblem object.

b

Solver object. Not need if a is an ConservationProblem object.

Value

A RasterLayer-class, Spatial, or a numeric vector containing the solution depending on the argument to a.

Details

If a OptimizationProblem is supplied, then the solution is returned as a logical showing the status of each planning unit. If a ConservationProblem is supplied, then the solution is returned in the format of the cost data contained inside it:

RasterLayer-classthe solution is returned as a RasterLayer-class object. Cell values denote the decision.

Spatialthe solution is returned as a SpatialPolygonsDataFrame or SpatialLinesDataFrame with a "Solution" column containing the decision values.

Examples

# build minimal conservation problem
p <- problem(sim_pu_raster, sim_features) +
  minimum_set_objective() +
  relative_targets(0.1)

#> Error in UseMethod("problem"): no applicable method for 'problem' applied to an object of class "c('RasterBrick', 'Raster', 'RasterStackBrick', 'BasicRaster')"

# solve the problem
solve(p)

#> Error in solve(p): object 'p' not found