Problem solvers

These functions are used to specify the method for solving conservation planning problems.

gurobi_solver(...)

lpsymphony_solver(...)

rsymphony_solver(...)

default_solver()

Arguments

...

arguments passed to solver function. The available arguments depends on the solver specified. See below for details on valid arguments.

Details

Each solver uses a function from a different package for solving problems. As such the arguments available to customize the process of solving the problem vary depending on the solver used. Below is list showing the table with the name of the solver and a link to the function used to solve the problem.

gurobi

gurobi

rsymphony_solver

Rsymphony_solve_LP

lpsymphony_solver

symphony_solve_LP

Examples

# create basic problem and use defaults
p <- problem(cost=sim_pu_raster, features=sim_features) +
  minimium_set_objective() +
  relative_targets(0.1)

#> Error in UseMethod("problem"): no applicable method for 'problem' applied to an object of class "c('RasterBrick', 'Raster', 'RasterStackBrick', 'BasicRaster')"

# add lpsymphony solver with default parameters
p + lpsymphony_solver()

#> Error in eval(expr, envir, enclos): object 'p' not found

# add lpsymphony solver with custom parameters
p + lpsymphony_solver(gap=0.1, time_limit=100)

#> Error in eval(expr, envir, enclos): object 'p' not found

## Not run: ------------------------------------
# # add gurobi solver
# p + gurobi_solver(MIPGap=0.1, Presolve=2, TimeLimit=100)
# 
# # add Rsymhpony solver
# p + absolute_targets(gap=0.1, time_limit=100)
## ---------------------------------------------