Specify absolute targets

Specify targets expressed as the same values as the underlying feature data (ignoring any specified feature units). For example, setting a target of 10 requires that the solution secure a set of planning units for which their summed feature values are equal to or greater than 10. This function is designed to be used within add_auto_targets().

Usage

spec_absolute_targets(targets, ...)

Arguments

targets

numeric vector that specifies targets for each of the features. If a single value is specified, then all features are assigned the same target threshold.

...

not used.

Value

An object (Method) for specifying targets.

Mathematical formulation

This method involves setting target thresholds based on a pre-specified value. To express this mathematically, we will define the following terminology. Let \(a\) the absolute target for a feature (per targets). Given this terminology, the target threshold (\(t\)) for a feature is calculated as follows. $$t = a$$

Data calculations

This function involves calculating targets based on the spatial extent of the features in x. As such, this function may need to make particular assumptions depending on the data used to specify the planning units and features in x. If possible, when building x with problem(), it is recommended to specify features as a terra::rast() object to ensure correctness. Although this function can be used to specify targets for features in other data formats, you must take care to ensure that the information specifying the amount of each feature are in units of 1 km2. To help ensure that you are aware of this limitation, this function will produce a warning to alert you when the features in x belong to such data formats. If you are confident that your data are in the correct units, then you can safely ignore this warning. Below we provide details for ensuring that the targets are calculated correctly.

x has terra::rast() or sf::st_sf() planning units and terra::rast() features

This function will automatically ensure that the targets are calculated correctly. Given this, it is recommended to specify features in either of these when building a problem().

x has sf::st_sf() or data.frame planning units and character features

Here, the argument to features of problem() is used to specify which columns of the planning unit data denote the amount of each feature present within each planning unit. All of these columns must contain values in units of 1 km km2. For example, if features = c("spp1", "spp2", "spp3"), then the columns "spp1", "spp2", and "spp3" of the planning unit denote data the amount of the first, second, and third features (respectively) present within each planning unit. If the column "spp1" has a value of 5 in the first row of the planning unit data, then this means the first planning unit contains 5 km km2 of habitat for the feature "spp1".

x has data.frame planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij must contain a data.frame with the columns "pu", "species", "amount". For a given row, the "pu" column contains an integer value denoting the identifier for the planning unit, the "species" column contains an integer value denoting the identifier for the feature, and "amount" contains a numericvalue denoting the amount of the the feature (per the"species"column) that is present in the planning unit (per"pu"column). Given this, the "amount" column must contain values that in units of 1 km \ifelse{html}{\out{km<sup>2</sup>}}{\eqn{km^2}}. For example, if the argument torijhas a row where the"pu"column has a value of 1, the"species"column has a value of 2, and the"amount"` column has a value of 3: then this means that the planning unit with an identifier of 1 contains 3 km km2 of the feature associated with the identifier of 2.

x has numeric or matrix planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij_matrix of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij_matrix must contain a matrix (or list of matrix objects) that contains numeric values in units of 1 km km2.

See also

Examples

# TODO