Specify targets based on interpolating area-based thresholds

Specify targets by interpolating them between area-based thresholds. Briefly, this method involves (i) setting target thresholds for rare features to a particular percentage threshold, (ii) setting target thresholds for common features to a particular percentage threshold, and (iii) interpolating target thresholds for features with spatial distributions that range between the those for the rare and common features. Additionally, features can (optionally) have their targets capped at a particular threshold. This method is especially useful for setting targets based on interpolation procedures when features have data in an area-based unit of measurement (e.g., km2). Note that this function is designed to be used within add_auto_targets() and add_group_targets().

Usage

spec_interp_area_targets(
  rare_area_threshold,
  rare_relative_target,
  rare_area_target,
  rare_method,
  common_area_threshold,
  common_relative_target,
  common_area_target,
  common_method,
  cap_area_target,
  interp_method,
  area_units
)

Arguments

rare_area_threshold

numeric value indicating the threshold area for identifying rare features. This value must be expressed in the same units as the feature data. In particular, features with a total spatial extent smaller than this value will be considered rare during the target setting calculations.

rare_relative_target

numeric value indicating the relative target for rare features. Note that this value must be a proportion between 0 and 1. For example, a value of 0.1 corresponds to 10%.

rare_area_target

numeric value denoting the area-based target for rare features. This value must be expressed in the same units as area_units. To avoid setting an area-based target for rare features, a missing (NA) value can be specified.

rare_method

character value indicating how the target for rare features should be calculated. Available options include "min" and "max". For example, a value of "max" means that the target for a rare features is calculated as the maximum based on rare_relative_target and rare_area_threshold.

common_area_threshold

numeric value indicating the threshold area for identifying common features. This value must be expressed in the same units as area_units. In particular, features with a total spatial extent greater than this value will be considered common during the target setting calculations.

common_relative_target

numeric value denoting the relative target for common features. Note that this value must be a proportion between 0 and 1. For example, a value of 0.1 corresponds to 10%.

common_area_target

numeric value denoting the area-based target for common features. This value must be expressed in the same units as area_units. To avoid setting an area-based target for common features, a missing (NA) value can be specified.

common_method

character value indicating how the target for common features should be calculated. Available options include "min" and "max". For example, a value of "max" means that the target for a common feature is calculated as the maximum based on common_relative_target and common_area_threshold.

cap_area_target

numeric value denoting the area-based target cap. This value must be expressed in the same units as area_units. In particular, all targets are clamped to this value during target setting calculations. To avoid setting a target cap, a missing (NA) value can be specified.

interp_method

character value denoting the interpolation method. Available options include "linear" for linear interpolation and "log10" for log-linear interpolation.

area_units

character value denoting the unit of measurement for the area-based arguments (e.g., "km^2", "ha", "acres"`).

Value

An object (Method) for specifying targets.

Details

This method has been applied to set target thresholds at global and national scales (e.g., Butchart et al. 2015; Rodrigues et al. 2004; Polak *et al.*2015). It is based on the rationale that species with a smaller geographic distribution are at a greater risk of extinction, and so require a larger percentage of their geographic distribution to be represented by a prioritization (Rodrigues et al. 2004). When using this method in a planning exercise, it is important to ensure that the threshold parameters reflect the stakeholder objectives. Additionally, the threshold parameters may need to calibrated based on the spatial extent of the planning region.

Mathematical formulation

This method provides a flexible approach for setting target thresholds based on an interpolation procedure and the spatial extent of the features. To express this mathematically, we will define the following terminology. Let \(f\) denote the total spatial extent of a feature (e.g., geographic range size), \(a\) the threshold for identifying rare features (per rare_area_threshold), \(b\) the relative targets for rare features (per rare_relative_target), \(c\) the area-based targets for rare features (per rare_area_target), \(d()\) the function for calculating targets for rare features as a maximum or minimum value (per rare_method). \(e\) the threshold for identifying common features (per common_area_threshold), \(g\) the relative targets for common features (per common_relative_target), \(h\) the area-based targets for common features (per common_area_target), \(i()\) the method for calculating targets for common features as a maximum or minimum value (per common_method). \(j\) the target cap (per cap_area_target). \(k()\) the interpolation method for features with a spatial distribution that is larger than a rare features and smaller than a common feature (per interp_method). In particular, \(j()\) is either a linear or log-linear interpolation procedure based on the thresholds for identifying rare and common features as well as the relative targets for rare and common features. Given this terminology, the target threshold (\(t\)) for a feature is calculated as follows.

If \(f < a\), then \(t = min(d(c, b * f), j)\).

If \(f > e\), then \(t = min(i(h, g * f), j)\).

If \(a <= f <= e\), then \(t = min(k(f, a, b, e, g), j)\).

Data calculations

This function involves calculating targets based on the spatial extent of the features in x. As such, this function may need to make particular assumptions depending on the data used to specify the planning units and features in x. If possible, when building x with problem(), it is recommended to specify features as a terra::rast() object to ensure correctness. Although this function can be used to specify targets for features in other data formats, you must take care to ensure that the information specifying the amount of each feature are in units of 1 km2. To help ensure that you are aware of this limitation, this function will produce a warning to alert you when the features in x belong to such data formats. If you are confident that your data are in the correct units, then you can safely ignore this warning. Below we provide details for ensuring that the targets are calculated correctly.

x has terra::rast() or sf::st_sf() planning units and terra::rast() features

This function will automatically ensure that the targets are calculated correctly. Given this, it is recommended to specify features in either of these when building a problem().

x has sf::st_sf() or data.frame planning units and character features

Here, the argument to features of problem() is used to specify which columns of the planning unit data denote the amount of each feature present within each planning unit. All of these columns must contain values in units of 1 km km2. For example, if features = c("spp1", "spp2", "spp3"), then the columns "spp1", "spp2", and "spp3" of the planning unit denote data the amount of the first, second, and third features (respectively) present within each planning unit. If the column "spp1" has a value of 5 in the first row of the planning unit data, then this means the first planning unit contains 5 km km2 of habitat for the feature "spp1".

x has data.frame planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij must contain a data.frame with the columns "pu", "species", "amount". For a given row, the "pu" column contains an integer value denoting the identifier for the planning unit, the "species" column contains an integer value denoting the identifier for the feature, and "amount" contains a numericvalue denoting the amount of the the feature (per the"species"column) that is present in the planning unit (per"pu"column). Given this, the "amount" column must contain values that in units of 1 km \ifelse{html}{\out{km<sup>2</sup>}}{\eqn{km^2}}. For example, if the argument torijhas a row where the"pu"column has a value of 1, the"species"column has a value of 2, and the"amount"` column has a value of 3: then this means that the planning unit with an identifier of 1 contains 3 km km2 of the feature associated with the identifier of 2.

x has numeric or matrix planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij_matrix of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij_matrix must contain a matrix (or list of matrix objects) that contains numeric values in units of 1 km km2.

References

Butchart SHM, Clarke M, Smith RJ, Sykes RE, Scharlemann JPW, Harfoot M, Buchanan GM, Angulo A, Balmford A, Bertzky B, Brooks TM, Carpenter KE, Comeros‐Raynal MT, Cornell J, Ficetola GF, Fishpool LDC, Fuller RA, Geldmann J, Harwell H, Hilton‐Taylor C, Hoffmann M, Joolia A, Joppa L, Kingston N, May I, Milam A, Polidoro B, Ralph G, Richman N, Rondinini C, Segan DB, Skolnik B, Spalding MD, Stuart SN, Symes A, Taylor J, Visconti P, Watson JEM, Wood L, Burgess ND (2015) Shortfalls and solutions for meeting national and global conservation area targets. Conservation Letters, 8: 329–337.

Polak T, Watson JEM, Fuller RA, Joseph LN, Martin TG, Possingham HP, Venter O, Carwardine J (2015) Efficient expansion of global protected areas requires simultaneous planning for species and ecosystems. Royal Society Open Science, 2: 150107.

Rodrigues ASL, Akçakaya HR, Andelman SJ, Bakarr MI, Boitani L, Brooks TM, Chanson JS, Fishpool LDC, Da Fonseca GAB, Gaston KJ, Hoffmann M, Marquet PA, Pilgrim JD, Pressey RL, Schipper J, Sechrest W, Stuart SN, Underhill LG, Waller RW, Watts MEJ, Yan X (2004) Global gap analysis: priority regions for expanding the global protected-area network. BioScience, 54: 1092–1100.

See also

Examples

# \dontrun{
# TODO
# }