Specify targets following Ward et al. (2025)

Specify targets based on the methodology outlined by Ward et al. (2025). Briefly, this method involves setting targets based the criteria for recognizing Critically Endangered species by the International Union for the Conservation of Nature (IUCN) Red List of Threatened Species (IUCN 2025). To help prevent widespread features from obscuring priorities, targets are capped following Butchart et al. (2015). This method was designed for species protection at national-scales. Note that this function is designed to be used within add_auto_targets() and add_group_targets().

Usage

spec_ward_targets(status = "CR", cap_area_target = 1e+06, area_units = "km^2")

Arguments

status

character value denoting the IUCN Red List threat status used for target setting. Available options include "CR" (Critically Endangered) , "EN" (Endangered), and "VU"(Vulnerable). Defaults to"CR"`.

cap_area_target

numeric value denoting the area-based target cap. To avoid setting a target cap, a missing (NA) value can be specified. Defaults to 1000000 (i.e., 1,000,000 km2).

area_units

character value denoting the unit of measurement for the area-based arguments. Defaults to "km^2" (i.e., km2).

Value

An object (Method) for specifying targets.

Details

This target setting method was designed to protect species in a national-scale prioritizations (Ward et al. 2025). Since it was designed for national-scale prioritizations, it may fail to identify meaningful priorities for prioritizations conducted at smaller geographic scales (e.g., national, state-level or county scales). For example, if this method is applied to smaller geographic scales, then the resulting prioritizations may select an overly large percentage of the study area, or be biased towards over-representing common and widespread species. As such, if you are working at smaller scales, it is recommended to set thresholds based on that criteria are appropriate to the spatial extent of the planning region. Please note that this function is provided as convenient method to set targets for problems with a single management zone, and cannot be used for those with multiple management zones.

Mathematical formulation

This method involves setting target thresholds based on assessment criteria from the IUCN Red List (IUCN 2025). It is based on the idea that a protected area system should ideally safeguard enough of a species' distribution to ensure that a species would, at worst, be classified under a particular threat status. In particular, this method considers criteria related to the size of a species' spatial distribution (i.e., Criterion B) and population size reduction (i.e., Criterion A). By default, it considers criteria for the Critically Endangered threat status and involves setting the target threshold for a species as 100,000 km2 (per subcriterion B1) or 20% (per subcriterion A2) of its spatial distribution (which ever value is larger). Additionally, following Butchart et al. (2015), a cap of 1,000,000 km2 is applied to target thresholds. By helping to ensure that species would – at a minimum – meet criteria for being recognized as Critically Endangered, this method aims to reduce chances that species will become extinct.

Data calculations

This function involves calculating targets based on the spatial extent of the features in x. As such, this function may need to make particular assumptions depending on the data used to specify the planning units and features in x. If possible, when building x with problem(), it is recommended to specify features as a terra::rast() object to ensure correctness. Although this function can be used to specify targets for features in other data formats, you must take care to ensure that the information specifying the amount of each feature are in units of 1 km2. To help ensure that you are aware of this limitation, this function will produce a warning to alert you when the features in x belong to such data formats. If you are confident that your data are in the correct units, then you can safely ignore this warning. Below we provide details for ensuring that the targets are calculated correctly.

x has terra::rast() or sf::st_sf() planning units and terra::rast() features

This function will automatically ensure that the targets are calculated correctly. Given this, it is recommended to specify features in either of these when building a problem().

x has sf::st_sf() or data.frame planning units and character features

Here, the argument to features of problem() is used to specify which columns of the planning unit data denote the amount of each feature present within each planning unit. All of these columns must contain values in units of 1 km km2. For example, if features = c("spp1", "spp2", "spp3"), then the columns "spp1", "spp2", and "spp3" of the planning unit denote data the amount of the first, second, and third features (respectively) present within each planning unit. If the column "spp1" has a value of 5 in the first row of the planning unit data, then this means the first planning unit contains 5 km km2 of habitat for the feature "spp1".

x has data.frame planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij must contain a data.frame with the columns "pu", "species", "amount". For a given row, the "pu" column contains an integer value denoting the identifier for the planning unit, the "species" column contains an integer value denoting the identifier for the feature, and "amount" contains a numericvalue denoting the amount of the the feature (per the"species"column) that is present in the planning unit (per"pu"column). Given this, the "amount" column must contain values that in units of 1 km \ifelse{html}{\out{km<sup>2</sup>}}{\eqn{km^2}}. For example, if the argument torijhas a row where the"pu"column has a value of 1, the"species"column has a value of 2, and the"amount"` column has a value of 3: then this means that the planning unit with an identifier of 1 contains 3 km km2 of the feature associated with the identifier of 2.

x has numeric or matrix planning units and data.frame features

The argument to features of problem() is used to specify a unique identifier and name for each of the features. Additionally, the argument to rij_matrix of problem() is used to specify the amount of each feature present within each planning unit. The argument to rij_matrix must contain a matrix (or list of matrix objects) that contains numeric values in units of 1 km km2.

References

Butchart SHM, Clarke M, Smith RJ, Sykes RE, Scharlemann JPW, Harfoot M, Buchanan GM, Angulo A, Balmford A, Bertzky B, Brooks TM, Carpenter KE, Comeros‐Raynal MT, Cornell J, Ficetola GF, Fishpool LDC, Fuller RA, Geldmann J, Harwell H, Hilton‐Taylor C, Hoffmann M, Joolia A, Joppa L, Kingston N, May I, Milam A, Polidoro B, Ralph G, Richman N, Rondinini C, Segan DB, Skolnik B, Spalding MD, Stuart SN, Symes A, Taylor J, Visconti P, Watson JEM, Wood L, Burgess ND (2015) Shortfalls and solutions for meeting national and global conservation area targets. Conservation Letters, 8: 329–337.

IUCN (2025) The IUCN Red List of Threatened Species. Version 2025-1. Available at https://www.iucnredlist.org. Accessed on 23 July 2025.

Ward M, Possingham HP, Wintle BA, Woinarski JCZ, Marsh JR, Chapple DG, Lintermans M, Scheele BC, Whiterod NS, Hoskin CJ, Aska B, Yong C, Tulloch A, Stewart R, Watson JEM (2025) The estimated cost of preventing extinction and progressing recovery for Australia's priority threatened species. Proceedings of the National Academy of Sciences, 122: e2414985122.

See also

Examples

#TODO

#TODO