Calculate representation level of features in planning units

Summarize a RasterStack of conservation features over a set of planning unit polygons. If the marxan package is installed, this function is a wrapper for calcPuVsSpeciesData from the marxan package (available at https://github.com/jeffreyhanson/marxan), otherwise a much slower implementation is used that doesn't require marxan.

summarize_features(pu, features, matrix = TRUE, sparse = TRUE)

Arguments

pu

SpatialPolygons* object; planning unit polygons.

features

RasterStack object; distribution of features over study area where each feature is a seperate layer in the RasterStack. Absences can be indicated by a cell value of zero or NA.

matrix

logical; whether to return a matrix (TRUE) or data frame (FALSE).

sparse

logical; whether to return a normal matrix or a sparse matrix as implemented by the slam package. Sparse matrices only explicitly store non-zero elements, which uses less memory.

Value

A matrix of feature representation levels within planning units where rows correspond to features and columns correspond to planning units. If matrix = FALSE a data frame is returned instead with columns corresponding to feature index, planning unit index, and amount within planning unit, repsectively.

Examples

# generate 4 feature distributions
e <- raster::extent(0, 100, 0, 100)
r <- raster::raster(e, nrows = 100, ncols = 100, vals = 1)
r_rare <- gaussian_field(r, range = 20, n = 2, prop = 0.1)
r_common <- gaussian_field(r, range = 20, n = 2, prop = 0.5)
r_features <- raster::stack(r_rare, r_common)
names(r_features) <- letters[1:4]

# create hexagonal grid
hex_grid <- make_grid(r, cell_area = 1000, type = "hexagonal")

# summarize
# spars matrix (default)
rij_sparse <- summarize_features(hex_grid, r_features)
rij_sparse

#> A 4x20 simple triplet matrix.

# normal matrix
rij_matrix <- summarize_features(hex_grid, r_features, sparse = FALSE)
dim(rij_matrix)

#> [1]  4 20

# data frame
rij_df <- summarize_features(hex_grid, r_features, matrix = FALSE)
head(rij_df)

#>   feature pu amount
#> 2       1  2     16
#> 3       1  3     70
#> 5       1  5    122
#> 6       1  6    123
#> 7       1  7      6
#> 9       1  9    105