Target weights

Targets are used to express the minimum amount or poportion of a feature's distribution required to preserve it in a protected area network. However, when using a maximum coverage problem with a pre-specified budget, it may not be possible to preserve all features. Target weights can be used to favour the representation of some features over other features when making decisions about how the budget should be allocated.

target_weights(x)

Arguments

x

numeric target weights. Higher values for a feature indicate that it is more desireable to represent the feature in a network. Weights cannot have negative values.

Value

Constraint object containing the weight data.

Details

Note that target weights can only be applied to a maximum coverage type of planning problem (eg. maximum_coverage_objective, phylogenetic_coverage_objective).

See also

targets.

Examples

# create problem
p <- problem(cost=sim_pu_raster, features=sim_features) +
  maximum_coverage_objective() +
  relative_targets(0.1)

#> Error in UseMethod("problem"): no applicable method for 'problem' applied to an object of class "c('RasterBrick', 'Raster', 'RasterStackBrick', 'BasicRaster')"

# create weights based on rarity (1/number occurrences)
w <- 1/raster::cellStats(sim_features, 'sum')

# add weight the targets in a problem
p <- p + target_weights(w)

#> Error in eval(expr, envir, enclos): object 'p' not found