Targets

After constructing a basic conservation problem that specifies the cost and feature data, it needs to have targets added to it. Targets are used to specify the minimum amount or proportion of a feature's distribution that needs to be protected. Thus targets ensure that each species is adequately represented in the reserve network.

relative_targets(x)

absolute_targets(x)

loglinear_targets(x)

Arguments

x

numeric vector.

Value

Constraint object containing the target data.

Details

Targets can be specified using one of the functions below.

relative_targets

Targets are expressed as a proportion (between 0 and 1) of the maximum level of representation in the study area. The argument to x should have a single value if all features have the same target. Otherwise, the vector should have a value for each feature. In this case, targets are assigned to features based on the their position in the argument to x and the feature when specifying the problem.

absolute_targets

Targets are expressed as the actual value that needs to be represented in the prioritisation. The argument to x is treated the same as for relative_targets.

loglinear_targets

Targets are expressed as a proportion (between 0 and 1) which is calculated using a log-linear equation and four tuning parameters (as used in XXX et al. XXX). The first tuning parameter specifies the first cut-off range size, and the second the second cut-off range size, the third argument specifies the target required for species with a range size equal to or less than the first cut-off range size, and the fourth argument specifies the target required for species with a range size equal to or greater than the required range size.

See also

objectives, problem, target_weights.

Examples

# create basic problem
p <- problem(cost=sim_pu_raster, features=sim_features) +
  minimium_set_objective()

#> Error in UseMethod("problem"): no applicable method for 'problem' applied to an object of class "c('RasterBrick', 'Raster', 'RasterStackBrick', 'BasicRaster')"

# add relative targets
p + relative_targets(0.1)

#> Error in eval(expr, envir, enclos): object 'p' not found

# add absolute targets
p + absolute_targets(0.1)

#> Error in eval(expr, envir, enclos): object 'p' not found

# add log-linear target
p + loglinear_targets(c(1, 10, 0.9, 0.2))

#> Error in eval(expr, envir, enclos): object 'p' not found