Miscellaneous parameter prototype

This prototype is used to represent a parameter that can be any object. Only experts should interact directly with this prototype.

Fields

$id

character identifier for parameter.

$name

character name of parameter.

$value

tibble::tibble() object.

$validator

list object containing a function that is used to validate changes to the parameter.

$widget

list object containing a function used to construct a shiny interface for modifying values.

Usage

x$print()

x$show()

x$validate(x)

x$get()

x$set(x)

x$reset()

x$render(...)

Arguments

x

object used to set a new parameter value.

...

arguments passed to $widget.

Details

print

print the object.

show

show the object.

validate

check if a proposed new parameter is valid.

get

extract the parameter value.

set

update the parameter value.

reset

update the parameter value to be the default value.

render

create a shiny::shiny() widget to modify parameter values.

See also