Optimization problem class

The OptimizationProblem class is used to represent an optimization problem. Data are stored in memory and accessed using an external pointer. Only experts should interact with this class directly.

Fields

$ptr

externalptr object.

Usage

x$print() x$show() x$repr() x$ncol() x$nrow() x$ncell() x$modelsense() x$vtype() x$obj() x$A() x$rhs() x$sense() x$lb() x$ub() x$number_of_planning_units() x$number_of_features() x$row_ids() x$col_ids() x$compressed_formulation()

Arguments

ptr

externalptr object.

Details

print

print the object.

show

show the object.

repr

character representation of object.

ncol

integer number of columns (variables) in model matrix.

nrow

integer number of rows (constraints) in model matrix.

ncell

integer number of cells in model matrix.

modelsense

character model sense.

vtype

character vector of variable types.

obj

numeric vector of objective function.

A

dgCMatrix-class model matrix

rhs

numeric vector of right-hand-side constraints.

sense

character vector of constraint senses.

lb

numeric vector of lower bounds for each decision variable.

ub

numeric vector of upper bounds for each decision variable.

number_of_features

integer number of features in the problem.

number_of_planning_units

integer number of planning units in the problem.

col_ids

character names describing each decision variable (column) in the model matrix.

row_ids

character names describing each constraint (row) in in the model matrix.

compressed_formulation

is the optimization problem formulated using a compressed version of the rij matrix?

shuffle_columns

randomly shuffle the columns in the problem. This should almost never be called manually and only should only be called after the optimization problem has been fully constructed.