Portfolio class

This class is used to represent portfolios used in optimization. Only experts should use the fields and methods for this class directly.

See also

Super class

prioritizr::ConservationModifier -> Portfolio

Methods

Public methods

Portfolio$run()

Portfolio$clone()

Inherited methods

prioritizr::ConservationModifier$calculate()

prioritizr::ConservationModifier$get_data()

prioritizr::ConservationModifier$get_internal()

prioritizr::ConservationModifier$print()

prioritizr::ConservationModifier$repr()

prioritizr::ConservationModifier$set_data()

prioritizr::ConservationModifier$set_internal()

prioritizr::ConservationModifier$show()

Method run()

Run the portfolio to generate solutions.

Usage

Portfolio$run(x, solver)

Arguments

x

optimization_problem() object.

solver

Solver object.

Returns

list of solutions.

Method clone()

The objects of this class are cloneable with this method.

Usage

Portfolio$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.