Add boundary penalties

Add penalties to a conservation problem to favor solutions that clump selected planning units together into contiguous reserves. Uses shared boundary length as a measure of connectivity, equivelant to the boundary length modifier (BLM) in Marxan. Boundary data is calculated automatically unless the planning units in x are stored in a data.frame, in which case boundary data must be added as a matrix or data.frame. This function can only be used for symmetric relationships between planning units; for asymmetric relationships use the add_connectivity_penalties function.

add_boundary_penalties(x, penalty, edge_factor, boundary_data = NULL)

Arguments

x
ConservationProblem-class object.

penalty
numeric penalty for exposed edges. This is equivalent to the boundary length modifier (BLM) parameter in Marxan.

edge_factor
numeric proportion to scale edges that do not have any neighboring planning units. For example, an edge factor of 0.5 is commonly used for planning units along the coast line.

boundary_data
A matrix or data.frame object showing the shared boundary lengths between planning units. If boundary_data is a matrix then each row and column denote a planning unit and cell values represent their shared boundary. If boundary_data is data.frame then it must have the columns "id1", "id2", and "boundary", where each row shows the shared boundary between two planning units (as per the standard Marxan input format). This argument is required if the planning units in x are stored in a data.frame object. Otherwise, it is optional and the boundary data will be automatically calculated. Note that the boundary_data must be denote symmetric relationships between planning units. If asymmetric relationships are required, use the add_connectivity_penalties function.

Value

ConservationProblem-class object.

See also

constraints, add_connectivity_penalties, penalties.

Examples

# create basic problem
p1 <- problem(sim_pu_raster, sim_features) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2) %>%
  add_default_solver()

#> Error in add_default_solver(.): no optimization problem solvers found on system.

# create problem with low boundary penalties
p2 <- p1 %>% add_boundary_penalties(5, 1)

#> Error in eval(lhs, parent, parent): object 'p1' not found

# create problem with high boundary penalties but outer edges receive
# half the penalty as inner edges
p3 <- p1 %>% add_boundary_penalties(50, 0.5)

#> Error in eval(lhs, parent, parent): object 'p1' not found

# solve problems
s <- stack(solve(p1), solve(p2), solve(p3))

#> Error in solve(p1): object 'p1' not found

# plot solutions
plot(s, main = c("basic solution", "small penalties", "high penalties"))

#> Error in plot(s, main = c("basic solution", "small penalties", "high penalties")): object 's' not found