Add neighbor constraints

Add constraints to a conservation problem to ensure that all selected planning units have at least a certain number of neighbors.

add_neighbor_constraints(x, k, ...)

Arguments

x
ConservationProblem-class object.

k
integer number of neighbors each selected planning units must have.

...
arguments passed to connected_matrix.

Value

ConservationProblem-class object with the constraint added to it.

See also

constraints for all the available constraints, and penalties.

Examples

# create basic problem
p1 <- problem(sim_pu_polygons, sim_features) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.2)

# create problem with constraints that require 1 neighbor
p2 <- p1 %>% add_neighbor_constraints(1)

# create problem with constraints that require 2 neighbors
p3 <- p1 %>% add_neighbor_constraints(2)

# create problem with constraints that require 3 neighbors
p4 <- p1 %>% add_neighbor_constraints(3)

# solve problems
s <- list(solve(p1), solve(p2), solve(p3), solve(p4))

#> Error in add_default_solver(a): no optimization problem solvers found on system.

# plot solutions
par(mfrow=c(2,2))

plot(s[[1]], main="basic solution")

#> Error in plot(s[[1]], main = "basic solution"): object 's' not found

plot(s[[1]][s[[1]]$solution==1, ], col = "darkgreen", add = TRUE)

#> Error in plot(s[[1]][s[[1]]$solution == 1, ], col = "darkgreen", add = TRUE): object 's' not found

plot(s[[2]], main="1 neighbor")

#> Error in plot(s[[2]], main = "1 neighbor"): object 's' not found

plot(s[[2]][s[[2]]$solution==1, ], col = "darkgreen", add = TRUE)

#> Error in plot(s[[2]][s[[2]]$solution == 1, ], col = "darkgreen", add = TRUE): object 's' not found

plot(s[[3]], main="2 neighbors")

#> Error in plot(s[[3]], main = "2 neighbors"): object 's' not found

plot(s[[3]][s[[3]]$solution==1, ], col = "darkgreen", add = TRUE)

#> Error in plot(s[[3]][s[[3]]$solution == 1, ], col = "darkgreen", add = TRUE): object 's' not found

plot(s[[4]], main="3 neighbors")

#> Error in plot(s[[4]], main = "3 neighbors"): object 's' not found

plot(s[[4]][s[[4]]$solution==1, ], col = "darkgreen", add = TRUE)

#> Error in plot(s[[4]][s[[4]]$solution == 1, ], col = "darkgreen", add = TRUE): object 's' not found