Matrix parameters

Create a parameter that represents a matrix object.

numeric_matrix_parameter(
  name,
  value,
  lower_limit = .Machine$double.xmin,
  upper_limit = .Machine$double.xmax,
  symmetric = FALSE
)

binary_matrix_parameter(name, value, symmetric = FALSE)

Arguments

name
character name of parameter.

value
matrix object.

lower_limit
numeric values denoting the minimum acceptable value in the matrix. Defaults to the smallest possible number on the system.

upper_limit
numeric values denoting the maximum acceptable value in the matrix. Defaults to the smallest possible number on the system.

symmetric
logical must the must be matrix be symmetric? Defaults to FALSE.

Value

MiscParameter object.

Examples

# create matrix
m <- matrix(runif(9), ncol = 3)
colnames(m) <- letters[1:3]
rownames(m) <- letters[1:3]

# create a numeric matrix parameter
p1 <- numeric_matrix_parameter("m", m)
print(p1) # print it
#> <environment: 0x563db5d60580>
#> attr(,"class")
#> [1] "MiscParameter" "MiscParameter" "Parameter"     "pproto"       
#> [5] "proto"         "environment"  
p1$get() # get value
#>           a         b         c
#> a 0.8334376 0.5892456 0.2387031
#> b 0.4846932 0.5758689 0.2372128
#> c 0.8543447 0.4674843 0.2312509
p1$id # get id
#> id: 4294027d-cc49-4e1f-9365-99a89812ddfc
p1$validate(m[, -1]) # check if parameter can be updated
#> [1] FALSE
#> attr(,"msg")
#> [1] "ncol(m) not equal to ncol(value)"
p1$set(m + 1) # set parameter to new values
p1$print() # print it again
#> [1] "m"

# create a binary matrix parameter
m <- matrix(round(runif(9)), ncol = 3)
colnames(m) <- letters[1:3]
rownames(m) <- letters[1:3]

# create a binary matrix parameter
p2 <- binary_matrix_parameter("m", m)
print(p2) # print it
#> <environment: 0x563db528f010>
#> attr(,"class")
#> [1] "MiscParameter" "MiscParameter" "Parameter"     "pproto"       
#> [5] "proto"         "environment"  
p2$get() # get value
#>   a b c
#> a 1 0 1
#> b 0 1 1
#> c 0 0 1
p2$id # get id
#> id: 30b60002-fed1-430a-bdb1-6c4fa6bd4a7f
p2$validate(m[, -1]) # check if parameter can be updated
#> [1] FALSE
#> attr(,"msg")
#> [1] "ncol(m) not equal to ncol(value)"
p2$set(m + 1) # set parameter to new values
p2$print() # print it again
#> [1] "m"