Conservation planning problem

Create a systematic conservation planning problem. This function is used to specify the basic data used in a spatial prioritization problem: the spatial distribution of the planning units and their costs, as well as the features (eg. species, ecosystems) that need to be conserved. After constructing this ConservationProblem-class object, it can be customized to meet specific goals using objectives, targets, constraints, and penalties.

The basic prioritizr workflow starts with formulating the problem, then adding objectives and targets, as well as constraints and penalties as needed. Alternative decision formats can be specified using the decisions functions, otherwise the package will default to binary selection of planning units. Lastly, the type of solving algorithm must be specified (see solvers). Once formulated, the problem is solved using the solve() function, which will return a RasterLayer-class, Spatial-class, or a numeric vector containing the solution depending on the ConservationProblem-class

problem(x, features, ...)

# S3 method for Raster
problem(x, features, ...)

# S3 method for Spatial
problem(x, features, cost_column = names(x)[1], ...)

# S3 method for data.frame
problem(x, features, rij, ...)

# S3 method for numeric
problem(x, features, rij_matrix, ...)

Arguments

x
Raster-class, SpatialPolygonsDataFrame-class, SpatialLinesDataFrame-class, or data.frame object, or numeric vector, specifying the planning units to use in the reserve design exercise and their corresponding cost. It may be desirable to exclude some planning units from the analysis, for example those outside the study area. To exclude planning units, set the cost for those raster cells to NA, or use the add_locked_out_constraint.

features
The correct argument for features depends on the input to x.

Raster-class, Spatial-class

Raster-class object showing the distribution of conservation features. Missing values (i.e. NA values) can be used to indicate the absence of a feature in a particular cell instead of explicitly setting these cells to zero.

data.frame or numeric

data.frame object containing information on the features. The argument to feature_data must follow the conventions used by Marxan. Each row corresponds to a different feature. It must also contain the following columns:

"id"

integer unique identifier for each feature These identifiers are used in the argument to rij.

"name"

character name for each feature.

"prop"

numeric relative target for each feature (optional).

"amount"

numeric absolute target for each feature (optional).

...
not used.

cost_column
character name or integer indicating the column in the attribute table of a Spatial-class object with the cost data.

rij
data.frame containing information on the amount of each feature in each planning unit. This argument is only used argument to x is a data.frame. Similar to features, the argument to rij must follow the conventions used by Marxan. It must contain the following columns:

"pu"

integer planning unit identifier.

"species"

integer feature identifier.

"amount"

numeric amount of the feature in the planning unit.

rij_matrix
matrix or dgCMatrix-class object specifying the amount of each feature (rows) within each planning unit (columns). Only used when x is a numeric vector of costs.

Value

A ConservationProblem-class object containing the basic data used to build a prioritization problem.

Details

A reserve design exercise starts by dividing the study region into planning units (typically square or hexagonal cells) and, for each planning unit, assigning values that quantify socioeconomic cost and conservation benefit for a set of conservation features. The cost can be the acquisition cost of the land, the cost of management, the opportunity cost of foregone commercial activities (e.g. from logging or agriculture), or simply the area. The conservation features are typically species (e.g. Clouded Leopard) or habitats (e.g. mangroves or cloud forest). The benefit that each feature derives from a planning unit can take a variety of forms, but is typically either occupancy (i.e. presence or absence) or area of occurrence within each planning unit. Finally, in some types of reserve design models, for each conservation feature, representation targets must be set, such as 20 extent of cloud forest or 10,000km^2 of Clouded Leopard habitat (see targets).

The goal of the reserve design exercise is then to optimize the trade-off between conservation benefit and socioeconomic cost, i.e. to get the most benefit for your limited conservation funds. In general, the goal of an optimization problem is to minimize an objective function over a set of decision variables, subject to a series of constraints. The decision variables are what we control, usually there is one binary variable for each planning unit specifying whether or not to protect that unit (but other approaches are available, see decisions). The constraints can be thought of as rules that need to be followed, for example, that the reserve must stay within a certain budget or meet the representation targets.

Integer linear programming (ILP) is the subset of optimization algorithms used in this package to solve reserve design problems. The general form of an ILP problem can be expressed in matrix notation as: $$\mathit{Minimize}\: \mathbf{c}^{\mathbf{T}}\mathbf{x} \: \mathit{subject\, to} \mathbf{Ax}\geq= or\leq \mathbf{b}$$

Where x is a vector of decision variables, c and b are vectors of known coefficients, and A is the constraint matrix. The final term specifies a series of structural constaints where relational operators for the constraint can be either \(\ge, =, or \le\) the coefficients. For example, in the minimum set cover problem, c would be a vector of costs for each planning unit, b a vector of targets for each conservation feature, the relational operator would be \(\ge\) for all features, and A would be the representation matrix with \(A_{ij}=r_{ij}\), the representation level of feature i in planning unit j.

See also

constraints, objectives, targets, decisions, solvers.

Examples

# create problem using raster planning unit data
p1 <- problem(sim_pu_raster, sim_features) %>%
 add_min_set_objective() %>%
 add_relative_targets(0.2) %>%
 add_binary_decisions()

# create problem using polygon planning unit data
p2 <- problem(sim_pu_polygons, sim_features) %>%
 add_min_set_objective() %>%
 add_relative_targets(0.2) %>%
 add_binary_decisions()

# create problem using line planning unit data
p3 <- problem(sim_pu_lines, sim_features) %>%
 add_min_set_objective() %>%
 add_relative_targets(0.2) %>%
 add_binary_decisions()

# create problem using point planning unit data
p4 <- problem(sim_pu_points, sim_features) %>%
 add_min_set_objective() %>%
 add_relative_targets(0.2) %>%
 add_binary_decisions()

# alternatively one can supply pre-processed, aspatial data
costs <- sim_pu_polygons$cost
features <- data.frame(id = 1:nlayers(sim_features),
                       name = names(sim_features))
rij_mat <- rij_matrix(sim_pu_polygons, sim_features)
p5 <- problem(costs, features, rij_matrix = rij_mat) %>%
 add_min_set_objective() %>%
 add_relative_targets(0.2) %>%
 add_binary_decisions()

# solve problems
s <- list(solve(p1), solve(p2), solve(p3), solve(p4), solve(p5))

#> Error in add_default_solver(a): no optimization problem solvers found on system.

# plot solutions
par(mfrow=c(2,2))
plot(s[[1]], main = "raster data")

#> Error in plot(s[[1]], main = "raster data"): object 's' not found

plot(s[[2]], main = "polygon data")

#> Error in plot(s[[2]], main = "polygon data"): object 's' not found

plot(s[[2]][s[[2]]$solution == 1, ], col = "darkgreen", add = TRUE)

#> Error in plot(s[[2]][s[[2]]$solution == 1, ], col = "darkgreen", add = TRUE): object 's' not found

plot(s[[3]], main = "line data")

#> Error in plot(s[[3]], main = "line data"): object 's' not found

lines(s[[3]][s[[3]]$solution == 1, ], col = "darkgreen", lwd = 2)

#> Error in lines(s[[3]][s[[3]]$solution == 1, ], col = "darkgreen", lwd = 2): object 's' not found

plot(s[[4]], main = "point data", pch = 19)

#> Error in plot(s[[4]], main = "point data", pch = 19): object 's' not found

points(s[[4]][s[[4]]$solution == 1, ], col = "darkgreen", cex = 2, pch = 19)

#> Error in points(s[[4]][s[[4]]$solution == 1, ], col = "darkgreen", cex = 2,     pch = 19): object 's' not found