rij

Generate a rij matrix using spatial data sets. The rij contains data on the amount of each feature in each planning unit.

Arguments

x

RasterLayer-class or Spatial-class object representing the planning units

y

Raster-class object representing the features

fun

function used to summarise values. Defaults to sum. Note that this only used when x is a SpatialPolygons-class or a SpatialLines-class object. This function must have an 'na.rm' argument.

na.rm

logical should missing values be omitted when performing calculations? Defaults to TRUE. Note this argument only has an effect when x is a SpatialPolygons-class or a SpatialLines-class and velox is FALSE.

velox

logical should the velox be used for geoprocessing? Defaults to TRUE if the package is installed. Note that this only used when x is a SpatialPolygons-class object.

...

not used.

Value

Matrix{dgTMatrix-class} object.

Details

The sparse matrix represents the spatial intersection between the planning units and the features. Rows correspond to planning units, and columns correspond to features. Values correspond to the amount of the feature in the planning unit. For example, the amount of the third species in the second planning unit would be contained in the cell in the third column and in the second column.

This function can take a long to run for big data sets. To reduce processing time, the set_number_of_threads function can be used to allocate more computational resources. Additionally, dealing with planning units represented with SpatialPolygonsDataFrame object, the velox package can be installed to reduce processing time.

Generally, processing Spatial-class data takes much longer to process then Raster-class data, and so it is recomended to use Raster-class data for planning units where possible.

See also

set_number_of_threads, velox.