Feature by planning unit matrix

Generate a matrix showing the amount of each feature in each planning unit (also known as an rij matrix).Each row corresponds to a different feature and each column corresponds to a different feature.

# S4 method for Raster,Raster
rij_matrix(x, y, ...)

# S4 method for Spatial,Raster
rij_matrix(x, y, ...)

Arguments

x
Raster-class or Spatial-class object representing the planning units.

y
Raster-class object representing the features.

...
additional arguments passed to fast_extract if argument to x inherits from a Spatial-class object.

Value

Matrix{dgCMatrix-class} object.

Details

The sparse matrix represents the spatial intersection between the planning units and the features. Rows correspond to planning units, and columns correspond to features. Values correspond to the amount of the feature in the planning unit. For example, the amount of the third species in the second planning unit would be contained in the cell in the third column and in the second column.

This function can take a long to run for big data sets. To reduce processing time, the set_number_of_threads function can be used to allocate more computational resources. Additionally, dealing with planning units represented with SpatialPolygonsDataFrame object, the velox package can be installed to reduce processing time.

Generally, processing Spatial-class data takes much longer to process then Raster-class data, and so it is recommended to use Raster-class data for planning units where possible.

See also

Examples

# load data
data(sim_pu_raster, sim_pu_polygons, sim_pu_zones_stack)

# create rij matrix using raster layer planning units
rij_raster <- rij_matrix(sim_pu_raster, sim_features)
print(rij_raster)

#> 5 x 90 sparse Matrix of class "dgCMatrix"
#>                                                                              
#> layer.1 0.6230919 0.6162216 0.6137453 0.6199624 0.6370438 0.6644234 0.6989494
#> layer.2 0.2778374 0.2943842 0.3166185 0.3413962 0.3647340 0.3827657 0.3928479
#> layer.3 0.7709779 0.7598676 0.7401120 0.7108536 0.6719204 0.6242170 0.5699056
#> layer.4 0.3410940 0.3975756 0.4558384 0.5120591 0.5621200 0.6023231 0.6298715
#> layer.5 0.4377535 0.4342554 0.4378609 0.4494131 0.4695732 0.4985266 0.5356873
#>                                                                              
#> layer.1 0.7359098 0.7706104 0.7996998 0.6858482 0.6808119 0.6790213 0.6844397
#> layer.2 0.3943315 0.3887735 0.3796238 0.2835019 0.3013247 0.3229453 0.3450114
#> layer.3 0.5121092 0.4541722 0.3988693 0.7873724 0.7803152 0.7642128 0.7381365
#> layer.4 0.6429162 0.6403052 0.6214165 0.2930375 0.3467998 0.4054523 0.4651304
#> layer.5 0.5794895 0.6274012 0.6762551 0.4675758 0.4600645 0.4596690 0.4676646
#>                                                                              
#> layer.1 0.6990305 0.7221749 0.7509363 0.7811599 0.8089150 0.7438902 0.7408126
#> layer.2 0.3636366 0.3753811 0.3782483 0.3723130 0.3598153 0.2862544 0.3052679
#> layer.3 0.7017190 0.6556210 0.6018581 0.5436197 0.4845071 0.7977861 0.7947781
#> layer.4 0.5209079 0.5678412 0.6019563 0.6206310 0.6223708 0.2602491 0.3100000
#> layer.5 0.4849662 0.5118109 0.5474671 0.5900668 0.6367107 0.4949354 0.4829775
#>                                                                              
#> layer.1 0.7396270 0.7437309 0.7548365 0.7724803 0.7943082 0.8170455 0.8376412
#> layer.2 0.3262044 0.3455575 0.3596247 0.3654688 0.3617898 0.3494113 0.3312313
#> layer.3 0.7824988 0.7601160 0.7272181 0.6842884 0.6331142 0.5767195 0.5186605
#> layer.4 0.3664171 0.4260395 0.4837777 0.5339954 0.5718218 0.5938925 0.5982894
#> layer.5 0.4781348 0.4821778 0.4963020 0.5207909 0.5547445 0.5959661 0.6411719
#>                                                                              
#> layer.1 0.8539994 0.7938114 0.7924808 0.7917329 0.7943076 0.8016644 0.8136098
#> layer.2 0.3116212 0.2863874 0.3064273 0.3266604 0.3434706 0.3533961 0.3540435
#> layer.3 0.4620455 0.8035417 0.8046419 0.7962895 0.7778932 0.7491517 0.7104864
#> layer.4 0.5841669 0.2421824 0.2874125 0.3397755 0.3963248 0.4522750 0.5019212
#> layer.5 0.6865899 0.5193616 0.5031221 0.4939220 0.4940363 0.5049327 0.5269192
#>                                                                              
#> layer.1 0.8285189 0.8440543 0.8579758 0.8686401 0.8339566 0.8340265 0.8335918
#> layer.2 0.3448288 0.3273092 0.3049608 0.2823980 0.2845693 0.3053672 0.3248867
#> layer.3 0.6634684 0.6108264 0.5558342 0.5013794 0.8057272 0.8110448 0.8066935
#> layer.4 0.5399928 0.5626118 0.5674554 0.5535030 0.2376677 0.2781951 0.3252848
#> layer.5 0.5589083 0.5983767 0.6416881 0.6848208 0.5411324 0.5213382 0.5083700
#>                                                                              
#> layer.1 0.8347624 0.8387518 0.8544114 0.8635984 0.8771618 0.8642933 0.8654516
#> layer.2 0.3394378 0.3458246 0.3286641 0.3074397 0.2583546 0.2817359 0.3029040
#> layer.3 0.7924316 0.7681943 0.6926466 0.6451522 0.5441737 0.8050861 0.8147492
#> layer.4 0.3764116 0.4273164 0.5074401 0.5274795 0.5146462 0.2456678 0.2812196
#> layer.5 0.5049636 0.5128230 0.5619659 0.5991425 0.6799487 0.5609025 0.5387355
#>                                                                              
#> layer.1 0.8653000 0.8654162 0.8667760 0.8696000 0.8734162 0.8773121 0.8802318
#> layer.2 0.3216605 0.3342739 0.3378123 0.3309326 0.3143474 0.2908090 0.2645076
#> layer.3 0.8144499 0.8043641 0.7847494 0.7562342 0.7201598 0.6786232 0.6340323
#> layer.4 0.3220592 0.3659594 0.4092420 0.4472944 0.4754793 0.4899747 0.4882349
#> layer.5 0.5230058 0.5168179 0.5220564 0.5389926 0.5661482 0.6004488 0.6377854
#>                                                                              
#> layer.1 0.8811753 0.8858077 0.8878633 0.8880513 0.8875396 0.8869644 0.8869952
#> layer.2 0.2400242 0.2789296 0.2999648 0.3178324 0.3287962 0.3210361 0.3025978
#> layer.3 0.5883783 0.8019348 0.8160655 0.8198468 0.8138975 0.7755126 0.7452436
#> layer.4 0.4692029 0.2652977 0.2952811 0.3288869 0.3641307 0.4263561 0.4455522
#> layer.5 0.6739256 0.5793741 0.5563225 0.5391462 0.5311859 0.5491093 0.5735531
#>                                                                              
#> layer.1 0.8866683 0.8852939 0.8821049 0.8998318 0.9033851 0.9026531 0.9011177
#> layer.2 0.2779733 0.2514503 0.2273927 0.2771435 0.3142078 0.3237263 0.3235317
#> layer.3 0.7098907 0.6714534 0.6314481 0.7961163 0.8227063 0.8208028 0.8101069
#> layer.4 0.4521548 0.4439332 0.4201729 0.2953050 0.3442162 0.3697036 0.3927471
#> layer.5 0.6045017 0.6377530 0.6691235 0.5970773 0.5576379 0.5491104 0.5511986
#>                                                                              
#> layer.1 0.8990353 0.8879524 0.8811592 0.9075585 0.9114092 0.9126903 0.9100761
#> layer.2 0.3130476 0.2432208 0.2201690 0.2772073 0.2961150 0.3114558 0.3183338
#> layer.3 0.7916873 0.7050115 0.6706865 0.7870026 0.8103953 0.8224209 0.8178684
#> layer.4 0.4102311 0.4004600 0.3714105 0.3333208 0.3495899 0.3659622 0.3928463
#> layer.5 0.5640395 0.6419204 0.6679737 0.6142783 0.5942938 0.5787437 0.5728523
#>                                                                              
#> layer.1 0.9068395 0.8965569 0.8890082 0.8792491 0.9097860 0.9148176 0.9169689
#> layer.2 0.3072526 0.2641941 0.2395820 0.2179862 0.2797196 0.2965165 0.3100377
#> layer.3 0.8039269 0.7598923 0.7328133 0.7036856 0.7735543 0.8017421 0.8180104
#> layer.4 0.3988856 0.3843406 0.3608068 0.3267069 0.3753569 0.3840860 0.3913383
#> layer.5 0.5844784 0.6277502 0.6518789 0.6724713 0.6309992 0.6147328 0.6021666
#>                                                                    
#> layer.1 0.9168402 0.9148075 0.9055836 0.8982756 0.8888513 0.8768529
#> layer.2 0.3168208 0.3147630 0.2853276 0.2629171 0.2401839 0.2204757
#> layer.3 0.8239604 0.8212071 0.7956766 0.7759476 0.7533783 0.7286484
#> layer.4 0.3961220 0.3969206 0.3792907 0.3578110 0.3272556 0.2889239
#> layer.5 0.5963748 0.5991050 0.6278240 0.6484574 0.6682495 0.6836491

# create rij matrix using polygon planning units
rij_polygons <- rij_matrix(sim_pu_polygons, sim_features)
print(rij_polygons)

#> 5 x 90 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 90 column names '1', '2', '3' ... ]]

#>                                                                              
#> layer.1 0.6230919 0.6162216 0.6137453 0.6199624 0.6370438 0.6644234 0.6989494
#> layer.2 0.2778374 0.2943842 0.3166185 0.3413962 0.3647340 0.3827657 0.3928479
#> layer.3 0.7709779 0.7598676 0.7401120 0.7108536 0.6719204 0.6242170 0.5699056
#> layer.4 0.3410940 0.3975756 0.4558384 0.5120591 0.5621200 0.6023231 0.6298715
#> layer.5 0.4377535 0.4342554 0.4378609 0.4494131 0.4695732 0.4985266 0.5356873
#>                                                                              
#> layer.1 0.7359098 0.7706104 0.7996998 0.6858482 0.6808119 0.6790213 0.6844397
#> layer.2 0.3943315 0.3887735 0.3796238 0.2835019 0.3013247 0.3229453 0.3450114
#> layer.3 0.5121092 0.4541722 0.3988693 0.7873724 0.7803152 0.7642128 0.7381365
#> layer.4 0.6429162 0.6403052 0.6214165 0.2930375 0.3467998 0.4054523 0.4651304
#> layer.5 0.5794895 0.6274012 0.6762551 0.4675758 0.4600645 0.4596690 0.4676646
#>                                                                              
#> layer.1 0.6990305 0.7221749 0.7509363 0.7811599 0.8089150 0.7438902 0.7408126
#> layer.2 0.3636366 0.3753811 0.3782483 0.3723130 0.3598153 0.2862544 0.3052679
#> layer.3 0.7017190 0.6556210 0.6018581 0.5436197 0.4845071 0.7977861 0.7947781
#> layer.4 0.5209079 0.5678412 0.6019563 0.6206310 0.6223708 0.2602491 0.3100000
#> layer.5 0.4849662 0.5118109 0.5474671 0.5900668 0.6367107 0.4949354 0.4829775
#>                                                                              
#> layer.1 0.7396270 0.7437309 0.7548365 0.7724803 0.7943082 0.8170455 0.8376412
#> layer.2 0.3262044 0.3455575 0.3596247 0.3654688 0.3617898 0.3494113 0.3312313
#> layer.3 0.7824988 0.7601160 0.7272181 0.6842884 0.6331142 0.5767195 0.5186605
#> layer.4 0.3664171 0.4260395 0.4837777 0.5339954 0.5718218 0.5938925 0.5982894
#> layer.5 0.4781348 0.4821778 0.4963020 0.5207909 0.5547445 0.5959661 0.6411719
#>                                                                              
#> layer.1 0.8539994 0.7938114 0.7924808 0.7917329 0.7943076 0.8016644 0.8136098
#> layer.2 0.3116212 0.2863874 0.3064273 0.3266604 0.3434706 0.3533961 0.3540435
#> layer.3 0.4620455 0.8035417 0.8046419 0.7962895 0.7778932 0.7491517 0.7104864
#> layer.4 0.5841669 0.2421824 0.2874125 0.3397755 0.3963248 0.4522750 0.5019212
#> layer.5 0.6865899 0.5193616 0.5031221 0.4939220 0.4940363 0.5049327 0.5269192
#>                                                                              
#> layer.1 0.8285189 0.8440543 0.8579758 0.8686401 0.8339566 0.8340265 0.8335918
#> layer.2 0.3448288 0.3273092 0.3049608 0.2823980 0.2845693 0.3053672 0.3248867
#> layer.3 0.6634684 0.6108264 0.5558342 0.5013794 0.8057272 0.8110448 0.8066935
#> layer.4 0.5399928 0.5626118 0.5674554 0.5535030 0.2376677 0.2781951 0.3252848
#> layer.5 0.5589083 0.5983767 0.6416881 0.6848208 0.5411324 0.5213382 0.5083700
#>                                                                              
#> layer.1 0.8347624 0.8387518 0.8544114 0.8635984 0.8771618 0.8642933 0.8654516
#> layer.2 0.3394378 0.3458246 0.3286641 0.3074397 0.2583546 0.2817359 0.3029040
#> layer.3 0.7924316 0.7681943 0.6926466 0.6451522 0.5441737 0.8050861 0.8147492
#> layer.4 0.3764116 0.4273164 0.5074401 0.5274795 0.5146462 0.2456678 0.2812196
#> layer.5 0.5049636 0.5128230 0.5619659 0.5991425 0.6799487 0.5609025 0.5387355
#>                                                                              
#> layer.1 0.8653000 0.8654162 0.8667760 0.8696000 0.8734162 0.8773121 0.8802318
#> layer.2 0.3216605 0.3342739 0.3378123 0.3309326 0.3143474 0.2908090 0.2645076
#> layer.3 0.8144499 0.8043641 0.7847494 0.7562342 0.7201598 0.6786232 0.6340323
#> layer.4 0.3220592 0.3659594 0.4092420 0.4472944 0.4754793 0.4899747 0.4882349
#> layer.5 0.5230058 0.5168179 0.5220564 0.5389926 0.5661482 0.6004488 0.6377854
#>                                                                              
#> layer.1 0.8811753 0.8858077 0.8878633 0.8880513 0.8875396 0.8869644 0.8869952
#> layer.2 0.2400242 0.2789296 0.2999648 0.3178324 0.3287962 0.3210361 0.3025978
#> layer.3 0.5883783 0.8019348 0.8160655 0.8198468 0.8138975 0.7755126 0.7452436
#> layer.4 0.4692029 0.2652977 0.2952811 0.3288869 0.3641307 0.4263561 0.4455522
#> layer.5 0.6739256 0.5793741 0.5563225 0.5391462 0.5311859 0.5491093 0.5735531
#>                                                                              
#> layer.1 0.8866683 0.8852939 0.8821049 0.8998318 0.9033851 0.9026531 0.9011177
#> layer.2 0.2779733 0.2514503 0.2273927 0.2771435 0.3142078 0.3237263 0.3235317
#> layer.3 0.7098907 0.6714534 0.6314481 0.7961163 0.8227063 0.8208028 0.8101069
#> layer.4 0.4521548 0.4439332 0.4201729 0.2953050 0.3442162 0.3697036 0.3927471
#> layer.5 0.6045017 0.6377530 0.6691235 0.5970773 0.5576379 0.5491104 0.5511986
#>                                                                              
#> layer.1 0.8990353 0.8879524 0.8811592 0.9075585 0.9114092 0.9126903 0.9100761
#> layer.2 0.3130476 0.2432208 0.2201690 0.2772073 0.2961150 0.3114558 0.3183338
#> layer.3 0.7916873 0.7050115 0.6706865 0.7870026 0.8103953 0.8224209 0.8178684
#> layer.4 0.4102311 0.4004600 0.3714105 0.3333208 0.3495899 0.3659622 0.3928463
#> layer.5 0.5640395 0.6419204 0.6679737 0.6142783 0.5942938 0.5787437 0.5728523
#>                                                                              
#> layer.1 0.9068395 0.8965569 0.8890082 0.8792491 0.9097860 0.9148176 0.9169689
#> layer.2 0.3072526 0.2641941 0.2395820 0.2179862 0.2797196 0.2965165 0.3100377
#> layer.3 0.8039269 0.7598923 0.7328133 0.7036856 0.7735543 0.8017421 0.8180104
#> layer.4 0.3988856 0.3843406 0.3608068 0.3267069 0.3753569 0.3840860 0.3913383
#> layer.5 0.5844784 0.6277502 0.6518789 0.6724713 0.6309992 0.6147328 0.6021666
#>                                                                    
#> layer.1 0.9168402 0.9148075 0.9055836 0.8982756 0.8888513 0.8768529
#> layer.2 0.3168208 0.3147630 0.2853276 0.2629171 0.2401839 0.2204757
#> layer.3 0.8239604 0.8212071 0.7956766 0.7759476 0.7533783 0.7286484
#> layer.4 0.3961220 0.3969206 0.3792907 0.3578110 0.3272556 0.2889239
#> layer.5 0.5963748 0.5991050 0.6278240 0.6484574 0.6682495 0.6836491

# create rij matrix using raster stack planning units
rij_raster <- rij_matrix(sim_pu_zones_stack, sim_features)
print(rij_raster)

#> 5 x 90 sparse Matrix of class "dgCMatrix"
#>                                                                              
#> layer.1 0.6230919 0.6162216 0.6137453 0.6199624 0.6370438 0.6644234 0.7359098
#> layer.2 0.2778374 0.2943842 0.3166185 0.3413962 0.3647340 0.3827657 0.3943315
#> layer.3 0.7709779 0.7598676 0.7401120 0.7108536 0.6719204 0.6242170 0.5121092
#> layer.4 0.3410940 0.3975756 0.4558384 0.5120591 0.5621200 0.6023231 0.6429162
#> layer.5 0.4377535 0.4342554 0.4378609 0.4494131 0.4695732 0.4985266 0.5794895
#>                                                                              
#> layer.1 0.7706104 0.7996998 0.6858482 0.6808119 0.6790213 0.6844397 0.6990305
#> layer.2 0.3887735 0.3796238 0.2835019 0.3013247 0.3229453 0.3450114 0.3636366
#> layer.3 0.4541722 0.3988693 0.7873724 0.7803152 0.7642128 0.7381365 0.7017190
#> layer.4 0.6403052 0.6214165 0.2930375 0.3467998 0.4054523 0.4651304 0.5209079
#> layer.5 0.6274012 0.6762551 0.4675758 0.4600645 0.4596690 0.4676646 0.4849662
#>                                                                              
#> layer.1 0.7221749 0.7509363 0.8089150 0.8315229 0.7438902 0.7408126 0.7548365
#> layer.2 0.3753811 0.3782483 0.3598153 0.3447226 0.2862544 0.3052679 0.3596247
#> layer.3 0.6556210 0.6018581 0.4845071 0.4275947 0.7977861 0.7947781 0.7272181
#> layer.4 0.5678412 0.6019563 0.6223708 0.6064703 0.2602491 0.3100000 0.4837777
#> layer.5 0.5118109 0.5474671 0.6367107 0.6839297 0.4949354 0.4829775 0.4963020
#>                                                                              
#> layer.1 0.7724803 0.7943082 0.8170455 0.8376412 0.8539994 0.7938114 0.7924808
#> layer.2 0.3654688 0.3617898 0.3494113 0.3312313 0.3116212 0.2863874 0.3064273
#> layer.3 0.6842884 0.6331142 0.5767195 0.5186605 0.4620455 0.8035417 0.8046419
#> layer.4 0.5339954 0.5718218 0.5938925 0.5982894 0.5841669 0.2421824 0.2874125
#> layer.5 0.5207909 0.5547445 0.5959661 0.6411719 0.6865899 0.5193616 0.5031221
#>                                                                              
#> layer.1 0.7917329 0.7943076 0.8016644 0.8136098 0.8285189 0.8440543 0.8579758
#> layer.2 0.3266604 0.3434706 0.3533961 0.3540435 0.3448288 0.3273092 0.3049608
#> layer.3 0.7962895 0.7778932 0.7491517 0.7104864 0.6634684 0.6108264 0.5558342
#> layer.4 0.3397755 0.3963248 0.4522750 0.5019212 0.5399928 0.5626118 0.5674554
#> layer.5 0.4939220 0.4940363 0.5049327 0.5269192 0.5589083 0.5983767 0.6416881
#>                                                                              
#> layer.1 0.8686401 0.8339566 0.8340265 0.8335918 0.8387518 0.8456213 0.8635984
#> layer.2 0.2823980 0.2845693 0.3053672 0.3248867 0.3458246 0.3421985 0.3074397
#> layer.3 0.5013794 0.8057272 0.8110448 0.8066935 0.7681943 0.7344600 0.6451522
#> layer.4 0.5535030 0.2376677 0.2781951 0.3252848 0.4273164 0.4727082 0.5274795
#> layer.5 0.6848208 0.5411324 0.5213382 0.5083700 0.5128230 0.5322519 0.5991425
#>                                                                              
#> layer.1 0.8716116 0.8771618 0.8642933 0.8654516 0.8653000 0.8654162 0.8667760
#> layer.2 0.2824439 0.2583546 0.2817359 0.3029040 0.3216605 0.3342739 0.3378123
#> layer.3 0.5948342 0.5441737 0.8050861 0.8147492 0.8144499 0.8043641 0.7847494
#> layer.4 0.5302593 0.5146462 0.2456678 0.2812196 0.3220592 0.3659594 0.4092420
#> layer.5 0.6398633 0.6799487 0.5609025 0.5387355 0.5230058 0.5168179 0.5220564
#>                                                                              
#> layer.1 0.8696000 0.8734162 0.8802318 0.8858077 0.8878633 0.8880513 0.8875396
#> layer.2 0.3309326 0.3143474 0.2645076 0.2789296 0.2999648 0.3178324 0.3287962
#> layer.3 0.7562342 0.7201598 0.6340323 0.8019348 0.8160655 0.8198468 0.8138975
#> layer.4 0.4472944 0.4754793 0.4882349 0.2652977 0.2952811 0.3288869 0.3641307
#> layer.5 0.5389926 0.5661482 0.6377854 0.5793741 0.5563225 0.5391462 0.5311859
#>                                                                              
#> layer.1 0.8870858 0.8869644 0.8869952 0.8866683 0.8852939 0.8821049 0.8998318
#> layer.2 0.3301682 0.3210361 0.3025978 0.2779733 0.2514503 0.2273927 0.2771435
#> layer.3 0.7988407 0.7755126 0.7452436 0.7098907 0.6714534 0.6314481 0.7961163
#> layer.4 0.3979044 0.4263561 0.4455522 0.4521548 0.4439332 0.4201729 0.2953050
#> layer.5 0.5344353 0.5491093 0.5735531 0.6045017 0.6377530 0.6691235 0.5970773
#>                                                                              
#> layer.1 0.9027376 0.9033851 0.9026531 0.9011177 0.8990353 0.8963638 0.8928215
#> layer.2 0.2974478 0.3142078 0.3237263 0.3235317 0.3130476 0.2938084 0.2691237
#> layer.3 0.8148241 0.8227063 0.8208028 0.8101069 0.7916873 0.7669104 0.7374611
#> layer.4 0.3188450 0.3442162 0.3697036 0.3927471 0.4102311 0.4189716 0.4162717
#> layer.5 0.5747657 0.5576379 0.5491104 0.5511986 0.5640395 0.5858312 0.6131908
#>                                                                              
#> layer.1 0.8879524 0.8811592 0.9075585 0.9114092 0.9126903 0.9120971 0.9100761
#> layer.2 0.2432208 0.2201690 0.2772073 0.2961150 0.3114558 0.3196221 0.3183338
#> layer.3 0.7050115 0.6706865 0.7870026 0.8103953 0.8224209 0.8244671 0.8178684
#> layer.4 0.4004600 0.3714105 0.3333208 0.3495899 0.3659622 0.3810576 0.3928463
#> layer.5 0.6419204 0.6679737 0.6142783 0.5942938 0.5787437 0.5709514 0.5728523
#>                                                                              
#> layer.1 0.9068395 0.9023909 0.8965569 0.8890082 0.8792491 0.9097860 0.9148176
#> layer.2 0.3072526 0.2880962 0.2641941 0.2395820 0.2179862 0.2797196 0.2965165
#> layer.3 0.8039269 0.7840762 0.7598923 0.7328133 0.7036856 0.7735543 0.8017421
#> layer.4 0.3988856 0.3967128 0.3843406 0.3608068 0.3267069 0.3753569 0.3840860
#> layer.5 0.5844784 0.6039177 0.6277502 0.6518789 0.6724713 0.6309992 0.6147328
#>                                                                    
#> layer.1 0.9169689 0.9148075 0.9110507 0.9055836 0.8888513 0.8768529
#> layer.2 0.3100377 0.3147630 0.3036771 0.2853276 0.2401839 0.2204757
#> layer.3 0.8180104 0.8212071 0.8112716 0.7956766 0.7533783 0.7286484
#> layer.4 0.3913383 0.3969206 0.3919052 0.3792907 0.3272556 0.2889239
#> layer.5 0.6021666 0.5991050 0.6102497 0.6278240 0.6682495 0.6836491