Feature by planning unit matrix

Generate a matrix showing the amount of each feature in each planning unit (also known as an rij matrix).

Usage

rij_matrix(x, y, ...)

# S4 method for Raster,Raster
rij_matrix(x, y, ...)

# S4 method for SpatRaster,SpatRaster
rij_matrix(x, y, memory, ...)

# S4 method for Spatial,Raster
rij_matrix(x, y, fun, ...)

# S4 method for sf,Raster
rij_matrix(x, y, fun, ...)

# S4 method for sf,SpatRaster
rij_matrix(x, y, fun, ...)

Arguments

x

terra::rast() or sf::sf() object representing planning units.

y

terra::rast() object.

...

not used.

memory

logical should calculations be performed using a method that prioritizes reduced memory consumption over speed? If TRUE, then calculations are performed using a method that reduces memory consumption, but can take a long time to complete. If FALSE, then calculations are performed using a method that reduces run time, but will fail when insufficient memory is available. Defaults to NA, such that calculations are automatically performed using the best method given available memory and dataset sizes. Note that this parameter can only be used when the arguments to x and y are both terra::rast() objects.

fun

character for summarizing values inside each planning unit. This parameter is only used when the argument to x is a sf::sf() object. Defaults to "sum".

Value

A dgCMatrix sparse matrix object. The sparse matrix represents the spatial intersection between the planning units and the features. Rows correspond to features, and columns correspond to planning units. Values correspond to the amount (or presence/absence) of the feature in the planning unit. For example, the amount of the third species in the second planning unit would be stored in the third column and second row.

Details

Generally, processing sf::st_sf() data takes much longer to process than terra::rast() data. As such, it is recommended to use terra::rast() data for planning units where possible. The performance of this function for large terra::rast() datasets can be improved by increasing the GDAL cache size. The default cache size is 25 MB. For example, the following code can be used to set the cache size to 4 GB.

terra::gdalCache(size = 4000)

Examples

# \dontrun{
# load data
sim_pu_raster <- get_sim_pu_raster()
sim_pu_polygons <- get_sim_pu_polygons()
sim_zones_pu_raster <- get_sim_zones_pu_raster()
sim_features <- get_sim_features()

# create rij matrix using raster layer planning units
rij_raster <- rij_matrix(sim_pu_raster, sim_features)
print(rij_raster)
#> 5 x 90 sparse Matrix of class "dgCMatrix"
#>                                                                                
#> feature_1 0.7150548 0.6990429 0.6859317 0.6783193 0.6782107 0.6863253 0.7017231
#> feature_2 0.2900901 0.3052216 0.3266036 0.3510977 0.3750470 0.3950113 0.4084749
#> feature_3 0.8178213 0.8064534 0.7852441 0.7541320 0.7131559 0.6628435 0.6046203
#> feature_4 0.2199663 0.2713800 0.3296247 0.3900085 0.4468281 0.4946333 0.5292249
#> feature_5 0.4533809 0.4413639 0.4343547 0.4355628 0.4474871 0.4714105 0.5068508
#>                                                                                
#> feature_1 0.7219284 0.7435505 0.7631509 0.7505900 0.7357790 0.7233151 0.7156627
#> feature_2 0.4142271 0.4123446 0.4038949 0.2800841 0.2967547 0.3181723 0.3409791
#> feature_3 0.5411391 0.4763017 0.4147364 0.8275983 0.8194545 0.8013062 0.7730504
#> feature_4 0.5479991 0.5498153 0.5348758 0.2208512 0.2738525 0.3350935 0.3998241
#> feature_5 0.5510725 0.5991256 0.6448475 0.4741697 0.4569167 0.4453572 0.4428185
#>                                                                                
#> feature_1 0.7147517 0.7212831 0.7343615 0.7516415 0.7699543 0.7897989 0.7772875
#> feature_2 0.3615030 0.3765606 0.3841341 0.3836641 0.3759393 0.2716298 0.2901825
#> feature_3 0.7345400 0.6860553 0.6288231 0.5654560 0.5000330 0.8346555 0.8298830
#> feature_4 0.4618047 0.5148343 0.5541170 0.5767631 0.5815260 0.2249796 0.2776009
#> feature_5 0.4517499 0.4732935 0.5068285 0.5495481 0.5965013 0.5022737 0.4802429
#>                                                                                
#> feature_1 0.7661407 0.7585160 0.7561970 0.7599327 0.7690603 0.7816119 0.7948577
#> feature_2 0.3119889 0.3334326 0.3508138 0.3612252 0.3631965 0.3568911 0.3438731
#> feature_3 0.8151305 0.7903745 0.7553823 0.7102172 0.6558247 0.5945687 0.5304188
#> feature_4 0.3390011 0.4047008 0.4684589 0.5238040 0.5655910 0.5906737 0.5976492
#> feature_5 0.4646212 0.4587578 0.4649526 0.4840959 0.5153215 0.5556964 0.6003282
#>                                                                                
#> feature_1 0.8059966 0.8286461 0.8189873 0.8095434 0.8020377 0.7980313 0.7983423
#> feature_2 0.3265679 0.2650109 0.2854743 0.3077757 0.3280176 0.3424767 0.3485511
#> feature_3 0.4684656 0.8381119 0.8367193 0.8254115 0.8043530 0.7733527 0.7323690
#> feature_4 0.5863706 0.2330116 0.2831343 0.3416658 0.4047077 0.4665769 0.5210694
#> feature_5 0.6432595 0.5353515 0.5095593 0.4907793 0.4822721 0.4860960 0.5028092
#>                                                                                
#> feature_1 0.8026563 0.8095316 0.8167838 0.8220091 0.8633036 0.8564520 0.8487799
#> feature_2 0.3453627 0.3338433 0.3163452 0.2959188 0.2603208 0.2824472 0.3051395
#> feature_3 0.6821137 0.6246373 0.5636095 0.5039577 0.8370439 0.8390267 0.8310691
#> feature_4 0.5629582 0.5888172 0.5969403 0.5869119 0.2456164 0.2910969 0.3438600
#> feature_5 0.5312405 0.5683321 0.6093528 0.6487381 0.5700379 0.5418938 0.5211593
#>                                                                                
#> feature_1 0.8415392 0.8360026 0.8325552 0.8337363 0.8339533 0.8912849 0.8867860
#> feature_2 0.3242198 0.3359588 0.3303278 0.3143945 0.2708325 0.2575975 0.2809245
#> feature_3 0.8136049 0.7866027 0.7045768 0.6519656 0.5398929 0.8305082 0.8359634
#> feature_4 0.4008068 0.4572360 0.5471036 0.5718452 0.5700791 0.2632827 0.3020677
#> feature_5 0.5108758 0.5127704 0.5521633 0.5850800 0.6559584 0.6024207 0.5734186
#>                                                                                
#> feature_1 0.8807106 0.8738915 0.8672825 0.8616340 0.8571536 0.8533099 0.8488533
#> feature_2 0.3037493 0.3216412 0.3308881 0.3295497 0.3179542 0.2985197 0.2749763
#> feature_3 0.8312740 0.8171942 0.7939453 0.7616869 0.7210739 0.6737961 0.6228802
#> feature_4 0.3465595 0.3945501 0.4425012 0.4859430 0.5203217 0.5418532 0.5479865
#> feature_5 0.5520171 0.5408934 0.5413573 0.5531961 0.5746993 0.6028225 0.6335756
#>                                                                                
#> feature_1 0.8419809 0.9117805 0.9090526 0.9042812 0.8979800 0.8833313 0.8758136
#> feature_2 0.2512416 0.2570026 0.2809260 0.3035274 0.3201754 0.3229680 0.3083414
#> feature_3 0.5724977 0.8175488 0.8267525 0.8253873 0.8145533 0.7665224 0.7305762
#> feature_4 0.5375278 0.2860727 0.3163391 0.3506471 0.3876576 0.4590604 0.4860971
#> feature_5 0.6626552 0.6287730 0.6001593 0.5792111 0.5681102 0.5773113 0.5951498
#>                                                                                
#> feature_1 0.8679237 0.8586633 0.8463772 0.9251764 0.9199524 0.9141710 0.9067838
#> feature_2 0.2863446 0.2611997 0.2370352 0.2589605 0.3047683 0.3200938 0.3251705
#> feature_3 0.6886563 0.6435894 0.5991429 0.7972733 0.8129275 0.8053969 0.7890073
#> feature_4 0.5023710 0.5050812 0.4927354 0.3133764 0.3566663 0.3816158 0.4070643
#> feature_5 0.6183284 0.6433893 0.6667633 0.6461805 0.5990580 0.5886663 0.5878728
#>                                                                                
#> feature_1 0.8981802 0.8643680 0.8473654 0.9322748 0.9317639 0.9287528 0.9159813
#> feature_2 0.3186414 0.2521341 0.2280966 0.2641903 0.2872474 0.3081174 0.3253427
#> feature_3 0.7645150 0.6570162 0.6186335 0.7690766 0.7873422 0.7936492 0.7768533
#> feature_4 0.4304725 0.4559778 0.4408970 0.3437657 0.3535038 0.3646749 0.3908502
#> feature_5 0.5957192 0.6486412 0.6667338 0.6528234 0.6269820 0.6089827 0.5991743
#>                                                                                
#> feature_1 0.9066665 0.8822224 0.8658993 0.8449602 0.9336696 0.9339561 0.9315392
#> feature_2 0.3170430 0.2740393 0.2478478 0.2243090 0.2736019 0.2951441 0.3143908
#> feature_3 0.7561399 0.6972504 0.6635963 0.6310227 0.7330271 0.7567351 0.7677462
#> feature_4 0.4031866 0.4135759 0.4061247 0.3879679 0.3750617 0.3743479 0.3741982
#> feature_5 0.6058227 0.6325923 0.6479832 0.6617020 0.6480033 0.6238689 0.6078230
#>                                                                      
#> feature_1 0.9265291 0.9190510 0.8968893 0.8817701 0.8629507 0.8388748
#> feature_2 0.3265712 0.3282949 0.2992975 0.2742827 0.2483226 0.2255383
#> feature_3 0.7679649 0.7589920 0.7195554 0.6927857 0.6644490 0.6371886
#> feature_4 0.3753241 0.3774461 0.3785580 0.3729115 0.3601426 0.3391285
#> feature_5 0.6002170 0.6002851 0.6166496 0.6288149 0.6410043 0.6515710

# create rij matrix using polygon planning units
rij_polygons <- rij_matrix(sim_pu_polygons, sim_features)
print(rij_polygons)
#> 5 x 90 sparse Matrix of class "dgCMatrix"
#>                                                                                
#> feature_1 0.7150548 0.6990429 0.6859317 0.6783193 0.6782107 0.6863253 0.7017231
#> feature_2 0.2900901 0.3052216 0.3266036 0.3510977 0.3750470 0.3950113 0.4084749
#> feature_3 0.8178213 0.8064534 0.7852441 0.7541320 0.7131559 0.6628435 0.6046203
#> feature_4 0.2199663 0.2713800 0.3296247 0.3900085 0.4468281 0.4946333 0.5292249
#> feature_5 0.4533809 0.4413639 0.4343547 0.4355628 0.4474871 0.4714105 0.5068508
#>                                                                                
#> feature_1 0.7219284 0.7435505 0.7631509 0.7505900 0.7357790 0.7233151 0.7156627
#> feature_2 0.4142271 0.4123446 0.4038949 0.2800841 0.2967547 0.3181723 0.3409791
#> feature_3 0.5411391 0.4763017 0.4147364 0.8275983 0.8194545 0.8013062 0.7730504
#> feature_4 0.5479991 0.5498153 0.5348758 0.2208512 0.2738525 0.3350935 0.3998241
#> feature_5 0.5510725 0.5991256 0.6448475 0.4741697 0.4569167 0.4453572 0.4428185
#>                                                                                
#> feature_1 0.7147517 0.7212831 0.7343615 0.7516415 0.7699543 0.7897989 0.7772875
#> feature_2 0.3615030 0.3765606 0.3841341 0.3836641 0.3759393 0.2716298 0.2901825
#> feature_3 0.7345400 0.6860553 0.6288231 0.5654560 0.5000330 0.8346555 0.8298830
#> feature_4 0.4618047 0.5148343 0.5541170 0.5767631 0.5815260 0.2249796 0.2776009
#> feature_5 0.4517499 0.4732935 0.5068285 0.5495481 0.5965013 0.5022737 0.4802429
#>                                                                                
#> feature_1 0.7661407 0.7585160 0.7561970 0.7599327 0.7690603 0.7816119 0.7948577
#> feature_2 0.3119889 0.3334326 0.3508138 0.3612252 0.3631965 0.3568911 0.3438731
#> feature_3 0.8151305 0.7903745 0.7553823 0.7102172 0.6558247 0.5945687 0.5304188
#> feature_4 0.3390011 0.4047008 0.4684589 0.5238040 0.5655910 0.5906737 0.5976492
#> feature_5 0.4646212 0.4587578 0.4649526 0.4840959 0.5153215 0.5556964 0.6003282
#>                                                                                
#> feature_1 0.8059966 0.8286461 0.8189873 0.8095434 0.8020377 0.7980313 0.7983423
#> feature_2 0.3265679 0.2650109 0.2854743 0.3077757 0.3280176 0.3424767 0.3485511
#> feature_3 0.4684656 0.8381119 0.8367193 0.8254115 0.8043530 0.7733527 0.7323690
#> feature_4 0.5863706 0.2330116 0.2831343 0.3416658 0.4047077 0.4665769 0.5210694
#> feature_5 0.6432595 0.5353515 0.5095593 0.4907793 0.4822721 0.4860960 0.5028092
#>                                                                                
#> feature_1 0.8026563 0.8095316 0.8167838 0.8220091 0.8633036 0.8564520 0.8487799
#> feature_2 0.3453627 0.3338433 0.3163452 0.2959188 0.2603208 0.2824472 0.3051395
#> feature_3 0.6821137 0.6246373 0.5636095 0.5039577 0.8370439 0.8390267 0.8310691
#> feature_4 0.5629582 0.5888172 0.5969403 0.5869119 0.2456164 0.2910969 0.3438600
#> feature_5 0.5312405 0.5683321 0.6093528 0.6487381 0.5700379 0.5418938 0.5211593
#>                                                                                
#> feature_1 0.8415392 0.8360026 0.8325552 0.8337363 0.8339533 0.8912849 0.8867860
#> feature_2 0.3242198 0.3359588 0.3303278 0.3143945 0.2708325 0.2575975 0.2809245
#> feature_3 0.8136049 0.7866027 0.7045768 0.6519656 0.5398929 0.8305082 0.8359634
#> feature_4 0.4008068 0.4572360 0.5471036 0.5718452 0.5700791 0.2632827 0.3020677
#> feature_5 0.5108758 0.5127704 0.5521633 0.5850800 0.6559584 0.6024207 0.5734186
#>                                                                                
#> feature_1 0.8807106 0.8738915 0.8672825 0.8616340 0.8571536 0.8533099 0.8488533
#> feature_2 0.3037493 0.3216412 0.3308881 0.3295497 0.3179542 0.2985197 0.2749763
#> feature_3 0.8312740 0.8171942 0.7939453 0.7616869 0.7210739 0.6737961 0.6228802
#> feature_4 0.3465595 0.3945501 0.4425012 0.4859430 0.5203217 0.5418532 0.5479865
#> feature_5 0.5520171 0.5408934 0.5413573 0.5531961 0.5746993 0.6028225 0.6335756
#>                                                                                
#> feature_1 0.8419809 0.9117805 0.9090526 0.9042812 0.8979800 0.8833313 0.8758136
#> feature_2 0.2512416 0.2570026 0.2809260 0.3035274 0.3201754 0.3229680 0.3083414
#> feature_3 0.5724977 0.8175488 0.8267525 0.8253873 0.8145533 0.7665224 0.7305762
#> feature_4 0.5375278 0.2860727 0.3163391 0.3506471 0.3876576 0.4590604 0.4860971
#> feature_5 0.6626552 0.6287730 0.6001593 0.5792111 0.5681102 0.5773113 0.5951498
#>                                                                                
#> feature_1 0.8679237 0.8586633 0.8463772 0.9251764 0.9199524 0.9141710 0.9067838
#> feature_2 0.2863446 0.2611997 0.2370352 0.2589605 0.3047683 0.3200938 0.3251705
#> feature_3 0.6886563 0.6435894 0.5991429 0.7972733 0.8129275 0.8053969 0.7890073
#> feature_4 0.5023710 0.5050812 0.4927354 0.3133764 0.3566663 0.3816158 0.4070643
#> feature_5 0.6183284 0.6433893 0.6667633 0.6461805 0.5990580 0.5886663 0.5878728
#>                                                                                
#> feature_1 0.8981802 0.8643680 0.8473654 0.9322748 0.9317639 0.9287528 0.9159813
#> feature_2 0.3186414 0.2521341 0.2280966 0.2641903 0.2872474 0.3081174 0.3253427
#> feature_3 0.7645150 0.6570162 0.6186335 0.7690766 0.7873422 0.7936492 0.7768533
#> feature_4 0.4304725 0.4559778 0.4408970 0.3437657 0.3535038 0.3646749 0.3908502
#> feature_5 0.5957192 0.6486412 0.6667338 0.6528234 0.6269820 0.6089827 0.5991743
#>                                                                                
#> feature_1 0.9066665 0.8822224 0.8658993 0.8449602 0.9336696 0.9339561 0.9315392
#> feature_2 0.3170430 0.2740393 0.2478478 0.2243090 0.2736019 0.2951441 0.3143908
#> feature_3 0.7561399 0.6972504 0.6635963 0.6310227 0.7330271 0.7567351 0.7677462
#> feature_4 0.4031866 0.4135759 0.4061247 0.3879679 0.3750617 0.3743479 0.3741982
#> feature_5 0.6058227 0.6325923 0.6479832 0.6617020 0.6480033 0.6238689 0.6078230
#>                                                                      
#> feature_1 0.9265291 0.9190510 0.8968893 0.8817701 0.8629507 0.8388748
#> feature_2 0.3265712 0.3282949 0.2992975 0.2742827 0.2483226 0.2255383
#> feature_3 0.7679649 0.7589920 0.7195554 0.6927857 0.6644490 0.6371886
#> feature_4 0.3753241 0.3774461 0.3785580 0.3729115 0.3601426 0.3391285
#> feature_5 0.6002170 0.6002851 0.6166496 0.6288149 0.6410043 0.6515710

# create rij matrix using raster planning units with multiple zones
rij_zones_raster <- rij_matrix(sim_zones_pu_raster, sim_features)
print(rij_zones_raster)
#> 5 x 90 sparse Matrix of class "dgCMatrix"
#>                                                                                
#> feature_1 0.7150548 0.6990429 0.6859317 0.6783193 0.6782107 0.6863253 0.7219284
#> feature_2 0.2900901 0.3052216 0.3266036 0.3510977 0.3750470 0.3950113 0.4142271
#> feature_3 0.8178213 0.8064534 0.7852441 0.7541320 0.7131559 0.6628435 0.5411391
#> feature_4 0.2199663 0.2713800 0.3296247 0.3900085 0.4468281 0.4946333 0.5479991
#> feature_5 0.4533809 0.4413639 0.4343547 0.4355628 0.4474871 0.4714105 0.5510725
#>                                                                                
#> feature_1 0.7435505 0.7631509 0.7505900 0.7357790 0.7233151 0.7156627 0.7147517
#> feature_2 0.4123446 0.4038949 0.2800841 0.2967547 0.3181723 0.3409791 0.3615030
#> feature_3 0.4763017 0.4147364 0.8275983 0.8194545 0.8013062 0.7730504 0.7345400
#> feature_4 0.5498153 0.5348758 0.2208512 0.2738525 0.3350935 0.3998241 0.4618047
#> feature_5 0.5991256 0.6448475 0.4741697 0.4569167 0.4453572 0.4428185 0.4517499
#>                                                                                
#> feature_1 0.7212831 0.7343615 0.7699543 0.7861227 0.7897989 0.7772875 0.7561970
#> feature_2 0.3765606 0.3841341 0.3759393 0.3627013 0.2716298 0.2901825 0.3508138
#> feature_3 0.6860553 0.6288231 0.5000330 0.4375452 0.8346555 0.8298830 0.7553823
#> feature_4 0.5148343 0.5541170 0.5815260 0.5684388 0.2249796 0.2776009 0.4684589
#> feature_5 0.4732935 0.5068285 0.5965013 0.6415349 0.5022737 0.4802429 0.4649526
#>                                                                                
#> feature_1 0.7599327 0.7690603 0.7816119 0.7948577 0.8059966 0.8286461 0.8189873
#> feature_2 0.3612252 0.3631965 0.3568911 0.3438731 0.3265679 0.2650109 0.2854743
#> feature_3 0.7102172 0.6558247 0.5945687 0.5304188 0.4684656 0.8381119 0.8367193
#> feature_4 0.5238040 0.5655910 0.5906737 0.5976492 0.5863706 0.2330116 0.2831343
#> feature_5 0.4840959 0.5153215 0.5556964 0.6003282 0.6432595 0.5353515 0.5095593
#>                                                                                
#> feature_1 0.8095434 0.8020377 0.7980313 0.7983423 0.8026563 0.8095316 0.8167838
#> feature_2 0.3077757 0.3280176 0.3424767 0.3485511 0.3453627 0.3338433 0.3163452
#> feature_3 0.8254115 0.8043530 0.7733527 0.7323690 0.6821137 0.6246373 0.5636095
#> feature_4 0.3416658 0.4047077 0.4665769 0.5210694 0.5629582 0.5888172 0.5969403
#> feature_5 0.4907793 0.4822721 0.4860960 0.5028092 0.5312405 0.5683321 0.6093528
#>                                                                                
#> feature_1 0.8220091 0.8633036 0.8564520 0.8487799 0.8360026 0.8330038 0.8337363
#> feature_2 0.2959188 0.2603208 0.2824472 0.3051395 0.3359588 0.3380832 0.3143945
#> feature_3 0.5039577 0.8370439 0.8390267 0.8310691 0.7866027 0.7500573 0.6519656
#> feature_4 0.5869119 0.2456164 0.2910969 0.3438600 0.4572360 0.5076754 0.5718452
#> feature_5 0.6487381 0.5700379 0.5418938 0.5211593 0.5127704 0.5270232 0.5850800
#>                                                                                
#> feature_1 0.8348882 0.8339533 0.8912849 0.8867860 0.8807106 0.8738915 0.8672825
#> feature_2 0.2933767 0.2708325 0.2575975 0.2809245 0.3037493 0.3216412 0.3308881
#> feature_3 0.5955431 0.5398929 0.8305082 0.8359634 0.8312740 0.8171942 0.7939453
#> feature_4 0.5797670 0.5700791 0.2632827 0.3020677 0.3465595 0.3945501 0.4425012
#> feature_5 0.6213430 0.6559584 0.6024207 0.5734186 0.5520171 0.5408934 0.5413573
#>                                                                                
#> feature_1 0.8616340 0.8571536 0.8488533 0.9117805 0.9090526 0.9042812 0.8979800
#> feature_2 0.3295497 0.3179542 0.2749763 0.2570026 0.2809260 0.3035274 0.3201754
#> feature_3 0.7616869 0.7210739 0.6228802 0.8175488 0.8267525 0.8253873 0.8145533
#> feature_4 0.4859430 0.5203217 0.5479865 0.2860727 0.3163391 0.3506471 0.3876576
#> feature_5 0.5531961 0.5746993 0.6335756 0.6287730 0.6001593 0.5792111 0.5681102
#>                                                                                
#> feature_1 0.8908018 0.8833313 0.8758136 0.8679237 0.8586633 0.8463772 0.9251764
#> feature_2 0.3271982 0.3229680 0.3083414 0.2863446 0.2611997 0.2370352 0.2589605
#> feature_3 0.7947903 0.7665224 0.7305762 0.6886563 0.6435894 0.5991429 0.7972733
#> feature_4 0.4249398 0.4590604 0.4860971 0.5023710 0.5050812 0.4927354 0.3133764
#> feature_5 0.5676758 0.5773113 0.5951498 0.6183284 0.6433893 0.6667633 0.6461805
#>                                                                                
#> feature_1 0.9237208 0.9199524 0.9141710 0.9067838 0.8981802 0.8885294 0.8775561
#> feature_2 0.2828033 0.3047683 0.3200938 0.3251705 0.3186414 0.3017897 0.2781030
#> feature_3 0.8107003 0.8129275 0.8053969 0.7890073 0.7645150 0.7330004 0.6962687
#> feature_4 0.3337335 0.3566663 0.3816158 0.4070643 0.4304725 0.4485815 0.4580145
#> feature_5 0.6187615 0.5990580 0.5886663 0.5878728 0.5957192 0.6102524 0.6288607
#>                                                                                
#> feature_1 0.8643680 0.8473654 0.9322748 0.9317639 0.9287528 0.9234146 0.9159813
#> feature_2 0.2521341 0.2280966 0.2641903 0.2872474 0.3081174 0.3219892 0.3253427
#> feature_3 0.6570162 0.6186335 0.7690766 0.7873422 0.7936492 0.7897301 0.7768533
#> feature_4 0.4559778 0.4408970 0.3437657 0.3535038 0.3646749 0.3773969 0.3908502
#> feature_5 0.6486412 0.6667338 0.6528234 0.6269820 0.6089827 0.5997976 0.5991743
#>                                                                                
#> feature_1 0.9066665 0.8955188 0.8822224 0.8658993 0.8449602 0.9336696 0.9339561
#> feature_2 0.3170430 0.2986799 0.2740393 0.2478478 0.2243090 0.2736019 0.2951441
#> feature_3 0.7561399 0.7289606 0.6972504 0.6635963 0.6310227 0.7330271 0.7567351
#> feature_4 0.4031866 0.4117466 0.4135759 0.4061247 0.3879679 0.3750617 0.3743479
#> feature_5 0.6058227 0.6177441 0.6325923 0.6479832 0.6617020 0.6480033 0.6238689
#>                                                                      
#> feature_1 0.9315392 0.9190510 0.9091912 0.8968893 0.8629507 0.8388748
#> feature_2 0.3143908 0.3282949 0.3186099 0.2992975 0.2483226 0.2255383
#> feature_3 0.7677462 0.7589920 0.7423053 0.7195554 0.6644490 0.6371886
#> feature_4 0.3741982 0.3774461 0.3792402 0.3785580 0.3601426 0.3391285
#> feature_5 0.6078230 0.6002851 0.6064383 0.6166496 0.6410043 0.6515710
# }