Run calculations
Source: R/run_calculations.R

Execute preliminary calculations in a conservation problem and store the results for later use. This function is useful when creating slightly different versions of the same conservation planning problem that involve the same pre-processing steps (e.g. calculating boundary data), because means that the same calculations will not be run multiple times.

run_calculations(x)

Arguments

x
ConservationProblem-class object

Value

Invisible TRUE indicating success.

Details

This function is used for the effect of modifying the input ConservationProblem-class object. As such, it does not return anything. To use this function with pipe operators, use the %T>% operator and not the %>% operator.

Examples

# Let us imagine a scenario where we wanted to understand the effect of
# setting different targets on our solution.

# create a conservation problem with no targets
p <- problem(sim_pu_raster, sim_features) %>%
     add_min_set_objective() %>%
     add_boundary_penalties(10, 0.5)

# create a copies of p and add targets
p1 <- p %>% add_relative_targets(0.1)
p2 <- p %>% add_relative_targets(0.2)
p3 <- p %>% add_relative_targets(0.3)

# now solve each of the different problems and record the time spent
# solving them
s1 <- system.time({solve(p1); solve(p2); solve(p3)})

#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [3e+00, 8e+00]
#> Found heuristic solution: objective 18523.196992
#> Found heuristic solution: objective 2352.9617505
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 1.942136e+03, 76 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1942.13608    0   11 2352.96175 1942.13608  17.5%     -    0s
#> H    0     0                    1999.6758952 1942.13608  2.88%     -    0s
#> 
#> Explored 1 nodes (76 simplex iterations) in 0.00 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 1999.68 2352.96 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 1.999675895189e+03, best bound 1.942136075861e+03, gap 2.8775%
#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [6e+00, 2e+01]
#> Found heuristic solution: objective 18523.196992
#> Found heuristic solution: objective 4772.1073492
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 3.924388e+03, 133 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 3924.38807    0   11 4772.10735 3924.38807  17.8%     -    0s
#> H    0     0                    4030.0635629 3924.38807  2.62%     -    0s
#> 
#> Explored 1 nodes (133 simplex iterations) in 0.01 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 4030.06 4772.11 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 4.030063562899e+03, best bound 3.924388065991e+03, gap 2.6222%
#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [9e+00, 2e+01]
#> Found heuristic solution: objective 18523.196992
#> Found heuristic solution: objective 6801.6615498
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 5.937186e+03, 158 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 5937.18606    0   20 6801.66155 5937.18606  12.7%     -    0s
#> H    0     0                    6073.3081531 5937.18606  2.24%     -    0s
#> 
#> Explored 1 nodes (158 simplex iterations) in 0.01 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 6073.31 6801.66 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 6.073308153120e+03, best bound 5.937186060308e+03, gap 2.2413%

# This approach is inefficient. Since these problems all share the same
# planning units it is actually performing the same calculations three times.
# To avoid this, we can use the "run_calculations" function before creating
# the copies. Normally, R runs the calculations just before solving the
# problem

# recreate a conservation problem with no targets and tell R run the
# preliminary calculations. Note how we use the \%T>\% operator here.
p <- problem(sim_pu_raster, sim_features) %>%
  add_min_set_objective() %>%
  add_boundary_penalties(10, 0.5) %T>%
  run_calculations()

# create a copies of p and add targets just like before
p1 <- p %>% add_relative_targets(0.1)
p2 <- p %>% add_relative_targets(0.2)
p3 <- p %>% add_relative_targets(0.3)

# solve each of the different problems and record the time spent
# solving them
s2 <- system.time({solve(p1); solve(p2); solve(p3)})

#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [3e+00, 8e+00]
#> Found heuristic solution: objective 18523.196992
#> Found heuristic solution: objective 2352.9617505
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 1.942136e+03, 76 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1942.13608    0   11 2352.96175 1942.13608  17.5%     -    0s
#> H    0     0                    1999.6758952 1942.13608  2.88%     -    0s
#> 
#> Explored 1 nodes (76 simplex iterations) in 0.00 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 1999.68 2352.96 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 1.999675895189e+03, best bound 1.942136075861e+03, gap 2.8775%
#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [6e+00, 2e+01]
#> Found heuristic solution: objective 18523.196992
#> Found heuristic solution: objective 4772.1073492
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 3.924388e+03, 133 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 3924.38807    0   11 4772.10735 3924.38807  17.8%     -    0s
#> H    0     0                    4030.0635629 3924.38807  2.62%     -    0s
#> 
#> Explored 1 nodes (133 simplex iterations) in 0.01 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 4030.06 4772.11 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 4.030063562899e+03, best bound 3.924388065991e+03, gap 2.6222%
#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [9e+00, 2e+01]
#> Found heuristic solution: objective 18523.196992
#> Found heuristic solution: objective 6801.6615498
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 5.937186e+03, 158 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 5937.18606    0   20 6801.66155 5937.18606  12.7%     -    0s
#> H    0     0                    6073.3081531 5937.18606  2.24%     -    0s
#> 
#> Explored 1 nodes (158 simplex iterations) in 0.01 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 6073.31 6801.66 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 6.073308153120e+03, best bound 5.937186060308e+03, gap 2.2413%

# now lets compare the times
print(s1) # time spent without running preliminary calculations

#>    user  system elapsed 
#>   0.594   0.226   0.831 

print(s2) # time spent after running preliminary calculations

#>    user  system elapsed 
#>   0.469   0.270   0.728 

# As we can see, we can save a lot of time by running the preliminary
# calculations before making copies of the problem with slightly
# different constraints.