Run calculations

Execute preliminary calculations in a conservation problem and store the results for later reuse. This function is useful when creating slightly different versions of the same conservation planning problem that involve the same constraints, because means that the same calculations will not be run multiple times.

run_calculations(x)

Arguments

x
ConservationProblem-class object

Value

None.

Details

This function is used for the effect of modifying the input ConservationProblem-class object. As such, it does not return anything. To use this function with pipe operators, use the %T>% operator and not the %>% operator.

See also

tee.

Examples

# Let us imagine a scenario where we wanted to understand the effect of
# setting different targets on our solution.

# create a conservation problem with no targets
p <- problem(sim_pu_raster, sim_features) %>%
  add_min_set_objective() %>%
  add_boundary_penalties(10, 0.5)

# create a copies of p and add targets
p1 <- p %>% add_relative_targets(0.1)
p2 <- p %>% add_relative_targets(0.2)
p3 <- p %>% add_relative_targets(0.3)

# now solve each of the different problems and record the time spent
# solving them
s1 <- system.time({solve(p1); solve(p2); solve(p3)})

#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [3e+00, 7e+00]
#> Found heuristic solution: objective 18523.2
#> Found heuristic solution: objective 2161.27
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 1.742900e+03, 34 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1742.89982    0   11 2161.26789 1742.89982  19.4%     -    0s
#> H    0     0                    1808.8782232 1742.89982  3.65%     -    0s
#> 
#> Explored 1 nodes (34 simplex iterations) in 0.00 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 1808.88 2161.27 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 1.808878223248e+03, best bound 1.742899817970e+03, gap 3.6475%
#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [6e+00, 1e+01]
#> Found heuristic solution: objective 18523.2
#> Found heuristic solution: objective 4155.27
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 3.520510e+03, 61 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 3520.50952    0   13 4155.27447 3520.50952  15.3%     -    0s
#> H    0     0                    3613.3501753 3520.50952  2.57%     -    0s
#> 
#> Explored 1 nodes (61 simplex iterations) in 0.00 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 3613.35 4155.27 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 3.613350175289e+03, best bound 3.520509518424e+03, gap 2.5694%
#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [8e+00, 2e+01]
#> Found heuristic solution: objective 18523.2
#> Found heuristic solution: objective 6193.03
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 5.323621e+03, 66 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 5323.62085    0   19 6193.03484 5323.62085  14.0%     -    0s
#> H    0     0                    5455.1404915 5323.62085  2.41%     -    0s
#> 
#> Explored 1 nodes (66 simplex iterations) in 0.01 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 5455.14 6193.03 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 5.455140491494e+03, best bound 5.323620845374e+03, gap 2.4109%

# This approach is inefficient. Since these problems all share the same
# planning units it is actually performing the same calculations three times.
# To avoid this, we can use the "run_calculations" function before creating
# the copies. Normally, R runs the calculations just before solving the
# problem

# recreate a conservation problem with no targets and tell R run the
# preliminary calculations. Note how we use the \%T>\% operator here.
p <- problem(sim_pu_raster, sim_features) %>%
  add_min_set_objective() %>%
  add_boundary_penalties(10, 0.5) %T>%
  run_calculations()

# create a copies of p and add targets just like before
p1 <- p %>% add_relative_targets(0.1)
p2 <- p %>% add_relative_targets(0.2)
p3 <- p %>% add_relative_targets(0.3)

# solve each of the different problems and record the time spent
# solving them
s2 <- system.time({solve(p1); solve(p2); solve(p3)})

#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [3e+00, 7e+00]
#> Found heuristic solution: objective 18523.2
#> Found heuristic solution: objective 2161.27
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 1.742900e+03, 34 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1742.89982    0   11 2161.26789 1742.89982  19.4%     -    0s
#> H    0     0                    1808.8782232 1742.89982  3.65%     -    0s
#> 
#> Explored 1 nodes (34 simplex iterations) in 0.00 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 1808.88 2161.27 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 1.808878223248e+03, best bound 1.742899817970e+03, gap 3.6475%
#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [6e+00, 1e+01]
#> Found heuristic solution: objective 18523.2
#> Found heuristic solution: objective 4155.27
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 3.520510e+03, 61 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 3520.50952    0   13 4155.27447 3520.50952  15.3%     -    0s
#> H    0     0                    3613.3501753 3520.50952  2.57%     -    0s
#> 
#> Explored 1 nodes (61 simplex iterations) in 0.01 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 3613.35 4155.27 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 3.613350175289e+03, best bound 3.520509518424e+03, gap 2.5694%
#> Optimize a model with 293 rows, 234 columns and 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [2e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [8e+00, 2e+01]
#> Found heuristic solution: objective 18523.2
#> Found heuristic solution: objective 6193.03
#> Presolve time: 0.00s
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> Variable types: 0 continuous, 234 integer (234 binary)
#> Presolved: 293 rows, 234 columns, 1026 nonzeros
#> 
#> 
#> Root relaxation: objective 5.323621e+03, 66 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 5323.62085    0   19 6193.03484 5323.62085  14.0%     -    0s
#> H    0     0                    5455.1404915 5323.62085  2.41%     -    0s
#> 
#> Explored 1 nodes (66 simplex iterations) in 0.01 seconds
#> Thread count was 1 (of 4 available processors)
#> 
#> Solution count 3: 5455.14 6193.03 18523.2 
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 5.455140491494e+03, best bound 5.323620845374e+03, gap 2.4109%

# now lets compare the times
print(s1) # time spent without running preliminary calculations

#>    user  system elapsed 
#>   0.502   0.182   0.671 

print(s2) # time spent after running preliminary calculations

#>    user  system elapsed 
#>   0.445   0.201   0.628 

# As we can see, we can save a lot of time by running the preliminary
# calculations before making copies of the problem with slightly
# different constraints.