Solve a problem

Solve a ConservationProblem-class or an OptimizationProblem-class.

Arguments

a
ConservationProblem-class or an OptimizationProblem-class object.

b
Solver-class object. Not used if a is an ConservationProblem-class object.

...
arguments passed to compile.

Value

A RasterLayer-class, Spatial-class, or a numeric vector containing the solution depending on the argument to a.

Details

If a OptimizationProblem-class is supplied, then the solution is returned as a logical showing the status of each planning unit. If a ConservationProblem-class is supplied, then the solution is returned in the format of the cost data contained inside it:

RasterLayer-classthe solution is returned as a RasterLayer-class object. Cell values denote the decision.

Spatial-classthe solution is returned as a SpatialPolygonsDataFrame or SpatialLinesDataFrame with a "Solution" column containing the decision values.

Examples

# build minimal conservation problem
p <- problem(sim_pu_raster, sim_features) %>%
  add_min_set_objective() %>%
  add_relative_targets(0.1) %>%
  add_binary_decisions()

# solve the problem
s <- solve(p)

#> Error in add_default_solver(a): no optimization problem solvers found on system.

# print the solution
print(s)

#> Error in print(s): object 's' not found

# plot solution
plot(s, main="solution")

#> Error in plot(s, main = "solution"): object 's' not found