Problem solvers

Specify the software and configurations used to solve a conservation planning problem. Below is a list of different solvers that can be added to a ConservationProblem-class object. By default the best available software currently installed on the system will be used.

Details

One of gurobi, Rsymphony, or lsymphony packages must be installed prior to solving.

default_solver

This solver uses the best software currently installed on the system.

add_gurobi_solver

Gurobi is a state-of-the-art commercial optimization software with an R package interface. It is by far the fastest of the solvers available in this package, however, it is also the only solver that is not freely available. That said, licenses are available to academics at no cost. The gurobi package is distributed with the Gurobi software suite. This solver uses the gurobi package to solve problems.

add_rsymphony_solver

SYMPHONY is an open-source integer programming solver that is part of the Computational Infrastructure for Operations Research (COIN-OR) project, an initiative to promote development of open-source tools for operations research (a field that includes linear programming). The Rsymphony package provides an interface to COIN-OR and is available on CRAN. This solver uses the Rsymphony package to solve problems.

add_lpsymphony_solver

The lpsymphony package provides a different interface to the COIN-OR software suite. Unlike the Rsymhpony package, the lpsymphony package is distributed through Bioconductor. On Windows and Mac, lpsymphony may be easier to install. This solver uses the lpsymphony package to solve.